% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reTest.R
\name{addTest}
\alias{addTest}
\title{Add tests to an occTest output}
\usage{
addTest(occTest_result, my_test)
}
\arguments{
\item{occTest_result}{data.frame or tibble. Output from an occTest project}

\item{my_test}{data.frame or tibble. New tests following the syntax of occTest (see Details)}
}
\value{
data frame with the new user-defined tests incorporated (see Details)
}
\description{
Reorganize an output from occTest columns to incorporate user defined tests.
}
\details{
my_test needs to follow the syntax of occTest testType_testMethod_test, testType_testMethod_value, testType_testMethod_comment 
the parameter my_test needs to have at least a logical column with the convention [testType]_[testMethod]_[test]
 where testTypes are the different tests types implemented in occTest
Note that the result has not have the testType_scores update. For that you need to use reTest.
}
\examples{
\donttest{
### THIS IS A CUT DOWN  EXAMPLE 
### visit vignetteXtra-occTest for more info
occTest_output <- readRDS (system.file('ext/output_occTest.rds',package = 'occTest'))
myNewTest <- data.frame (humanDetection_myInventedTest_test=sample (c(TRUE,FALSE),
replace = TRUE,
size = nrow(occTest_output)))
out_test_newAdded <- addTest(occTest_result = occTest_output, my_test = myNewTest)
}
}
