% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_alphaHulls.R
\name{ah2sp}
\alias{ah2sp}
\title{Convert Alpha Hull object into a shapefile}
\usage{
ah2sp(
  x,
  increment = 360,
  rnd = 10,
  proj4string = CRS(as.character(NA)),
  tol = 1e-04
)
}
\arguments{
\item{x}{an alpha hull object}

\item{increment}{numeric. Increments}

\item{rnd}{numeric. Decimal rounding}

\item{proj4string}{crs object with the spatial projectoinsprojection}

\item{tol}{numeric. tolerance}
}
\value{
a sp polygon object
}
\description{
Convert Alpha Hull object into a shapefile
}
\details{
Function written by Andrew Bevan, found on R-sig-Geo, and modified by Pascal Title
}
\examples{

}
\seealso{
Alpha hulls are created with \code{\link{ahull}}.
}
\author{
Pascal Title (original version), JM Serra-Diaz (modifications)
}
