% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internals_foreignPkgs.R
\name{biogeo_outliers}
\alias{biogeo_outliers}
\title{Centroid detection function}
\usage{
biogeo_outliers(rid, species, dups, ev)
}
\arguments{
\item{rid}{\emph{numeric}. Row identifier}

\item{species}{\emph{numeric}. Column of species names}

\item{dups}{\emph{numeric}. a column of zeros and ones, where ones indicate duplicates}

\item{ev}{\emph{numeric}. values of the environmental variable}
}
\value{
a matrix with the test outputs for outlier detection (1/0)
}
\description{
Calculates the outliers using the reverse jackknife procedure (see rjack) and boxplot statistics (using boxplot.stats). 
If the value lies 1.5 times beyond the length of the box in the boxplot then it is considered to be an outlier.
}
\details{
This is a method and functions first applied in the biogeo package
}
\examples{
rid<-1:20
species<-rep("Species A",20)
dups=rep(0,20)
ev<-c(rnorm(19,mean=20,sd=1),40)
a<-biogeo_outliers(rid, species, dups, ev)
}
\seealso{
Other analysis: 
\code{\link{.nearestcell3}()},
\code{\link{centroidDetection}()},
\code{\link{countryStatusRangeAnalysis}()},
\code{\link{duplicatesexcludeAnalysis}()},
\code{\link{geoEnvAccuracy}()},
\code{\link{humanDetection}()},
\code{\link{timeAccuracy}()}
}
\author{
Mark Robertson
}
\concept{analysis}
\keyword{internal}
