% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_filter.R
\name{buildCustomThresholds}
\alias{buildCustomThresholds}
\title{Build tables of user specified filters}
\usage{
buildCustomThresholds(df, level)
}
\arguments{
\item{df}{data.frame outputed from occurrenceClassify}

\item{level}{numeric Level of application of the filtering process.}
}
\value{
data.frame
}
\description{
Build tables of user specified filters
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{filter}
