% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_modifiedFunctions.R
\name{cc_round_occTest}
\alias{cc_round_occTest}
\title{Flag records with regular pattern interval}
\usage{
cc_round_occTest(
  x,
  lon = "decimallongitude",
  lat = "decimallatitude",
  ds = "dataset",
  T1 = 7,
  reg_out_thresh = 2,
  reg_dist_min = 0.1,
  reg_dist_max = 2,
  min_unique_ds_size = 4,
  graphs = F,
  test = "both",
  value = "flagged",
  verbose = TRUE
)
}
\description{
own version of coordinate cleaner cc_round
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (adapted from CoordinateCleaner)
}
\concept{}
\keyword{internal}
