% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_modifiedFunctions.R
\name{cc_urb_occProfileR}
\alias{cc_urb_occProfileR}
\title{Flag values in urban areas (based on coordinate cleaner but modified for occProfileR to deliver it quitely)}
\usage{
cc_urb_occProfileR(
  x,
  lon = "decimallongitude",
  lat = "decimallatitude",
  ref = NULL,
  value = "clean",
  verbose = F,
  outdir = output.dir
)
}
\arguments{
\item{x}{Data.frame of species occurrences}

\item{lon}{}

\item{lat}{}

\item{ref}{}

\item{value}{}

\item{verbose}{}
}
\value{

}
\description{
flags values for urban areas
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}

}
\references{

}
\seealso{

}
\author{
author of the coordinatecleaner package (Zizka)
}
\concept{}
\keyword{internal}
