% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{centroid_assessment}
\alias{centroid_assessment}
\title{Centroid assessment using BIEN methods}
\usage{
centroid_assessment(occurrences, centroid_data)
}
\value{
list
}
\description{
Centroid assessment using BIEN methods
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
<<<<<<< HEAD
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
=======
Brian Maitner
>>>>>>> 5e4e062df6954368ce7919fa66946aef731d6618
}
\concept{}
\keyword{internal}
