% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{countryStatusRangeAnalysis}
\alias{countryStatusRangeAnalysis}
\title{Range analysis}
\usage{
countryStatusRangeAnalysis(
  df,
  xf,
  yf,
  .ntv.ctry,
  .inv.ctry,
  .c.field,
  .points.crs,
  .countries.shapefile,
  cfsf,
  excludeUnknownRanges = FALSE,
  excludeNotmatchCountry = FALSE,
  doRangeAnalysis = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{\emph{data.frame} of species occurrences}

\item{xf}{\emph{character}. The field in df containing the x coordinates}

\item{yf}{\emph{character}. The field in df containing the y coordinates}

\item{.ntv.ctry}{\emph{character} Native country in ISO3 coding}

\item{.inv.ctry}{\emph{character} Invasive country in ISO3 coding}

\item{.c.field}{\emph{character} Country field in the species \emph{data.frame} (df)}

\item{.points.crs}{\emph{crs} for the occurrence data}

\item{.countries.shapefile}{spatialPolygonDataFrame of political divisions}

\item{cfsf}{\emph{character} Column name of the .country spatialPolygonDataFrame indicaing the country in ISO3 codeing.}

\item{excludeUnknownRanges}{\emph{logical}. Should records be filtered if located in countries outside .ntv.ctry or .inv.ctry? Defalut FALSE}

\item{excludeNotmatchCountry}{\emph{logical}. Should records be if the reported country is different than the locatoin country? . Default FALSE}

\item{doRangeAnalysis}{\emph{logical} Should range analysis be performed?}

\item{verbose}{\emph{logical} Print messages? Default FALSE}
}
\value{
a \emph{list} of two \emph{data.frames} with filtered occurrence records (out of native and/or alien country) and non-filtered records (inside native/alien country)
}
\description{
Identify and filter species records based on countries where species is considered native or alien.
}
\details{
It returns a \emph{list} with two elements: stay (records that are excluded/filtered), continue (records that are not excluded from the filtering process)
}
\examples{
library (occTest)
sp_file <- system.file('ext/exampleOccData.csv',package='occTest')
occ_species_example <- read.table(file = sp_file,header = TRUE,sep = ',',as.is = TRUE)
occ <- occ_species_example[1:10,c('MAPX','MAPY','COUNTRYRECORD')]
#load a high resolution coastlines
pol_ctry <- system.file('ext/countries.rds',package='occTest') |> readRDS()
countryStatusRangeAnalysis(df=occ,
                           xf = 'MAPX',yf = 'MAPY',
                           .ntv.ctry = 'ESP',.inv.ctry = 'FRA',
                           .c.field = 'COUNTRYRECORD',
                           .countries.shapefile = sf::st_as_sf(pol_ctry),
                           cfsf='ISO',
                           .points.crs = sf::st_crs ('EPSG:4326'))
}
\seealso{
Other analysis: 
\code{\link{.nearestcell3}()},
\code{\link{biogeo_outliers}()},
\code{\link{centroidDetection}()},
\code{\link{duplicatesexcludeAnalysis}()},
\code{\link{geoEnvAccuracy}()},
\code{\link{humanDetection}()},
\code{\link{timeAccuracy}()}
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{analysis}
\keyword{internal}
