% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_nativeStatusCtry.R
\name{ctryToIso3}
\alias{ctryToIso3}
\title{Convert country names to ISO3 codes}
\usage{
ctryToIso3(x)
}
\arguments{
\item{x}{character. country name}
}
\description{
using package countrycode
}
\details{
right now not implemented with fuzzy matching, but is case insensitive.
}
\examples{
\dontrun{
example<-"goes here"
}
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{internal}
