% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{.add.to.qfield}
\alias{.add.to.qfield}
\title{QUALITY GRADINGS}
\usage{
.add.to.qfield(x, qfield = "quality.comment", new.comment,
  separation.charachter = ";")
}
\arguments{
\item{qfield}{Field to add quality information in, default is "quality.comment"}

\item{new.comment}{New comment}

\item{separation.charachter}{Character to separate comments (?)}

\item{df}{Data.frame of species occurrences}
}
\value{
list
}
\description{
Assess record quality
}
\examples{
\dontrun{
example<-"goes here"
}


}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{internal}
