% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Checks.R
\name{.addmainfields2}
\alias{.addmainfields2}
\title{checking main fields}
\usage{
.addmainfields2(dat, species, verbose = F)
}
\arguments{
\item{dat}{A dataframe containing occurrence data for checking.}

\item{species}{character. Name of the species}
}
\value{
Original dataframe, dat.  Used primarily to generate warning messages.
}
\description{
Verify that all main data fields are correctly populated.
}
\examples{
\dontrun{
Example<-"goes here"
}

}
\seealso{
Other checks: \code{\link{.check.geospatial.data}},
  \code{\link{.checkdatastr2}}, \code{\link{.checkfields}},
  \code{\link{.status.tracker.and.escaping}}
}
\concept{checks}
