% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaHulls.R
\name{.ah2sf}
\alias{.ah2sf}
\title{Convert ahull object to sf object}
\usage{
.ah2sf(x, increment = 360, rnd = 10, crs = 4326, tol = 1e-04)
}
\arguments{
\item{x}{{ahull} object}
}
\description{
convert ahull object to sf object
}
\details{
Function written by Andrew Bevan, found on R-sig-Geo, and modified by Pascal Title.
Modified to support sf objects 17 Nov 2022
}
\keyword{internal}
