% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_modifiedFunctions.R
\name{.cc_outl_occTest}
\alias{.cc_outl_occTest}
\title{Identify geographic outliers based on methods from CoordinateCleaner package}
\usage{
.cc_outl_occTest(
  x,
  lon = "decimallongitude",
  lat = "decimallatitude",
  species = "species",
  method = "quantile",
  mltpl = 5,
  tdi = 1000,
  value = "clean",
  sampling_thresh = 0,
  verbose = TRUE,
  min_occs = 7,
  thinning = FALSE,
  thinning_res = 0.5
)
}
\description{
own version of coordinate cleaner geographic outliers
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{internal}
