% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-modifiedFunctions.R
\name{.cd_ddmm_occTest}
\alias{.cd_ddmm_occTest}
\title{Flag records with regular pattern interval}
\usage{
.cd_ddmm_occTest(
  x,
  lon = "decimallongitude",
  lat = "decimallatitude",
  ds = "dataset",
  pvalue = 0.025,
  diff = 1,
  mat_size = 1000,
  min_span = 2,
  value = "clean",
  verbose = TRUE,
  diagnostic = FALSE
)
}
\value{
a clean data.frame
}
\description{
own version of coordinate cleaner cc_round
}
\author{
A Zizka (original author) Josep M Serra-Diaz (adapted from CoordinateCleaner)
}
\keyword{internal}
