% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.check.geospatial.data}
\alias{.check.geospatial.data}
\title{Checks on the projection of the spatial data}
\usage{
.check.geospatial.data(list.geospatial.objects, verbose = FALSE)
}
\arguments{
\item{list.geospatial.objects}{A list of geospatial objects.Default list includes: 'countries.shapefile','r.env','r.dem','ras.hii','points.proj4string'}

\item{verbose}{logical. Print messages?}
}
\value{
None. Used to generate warning messages.
}
\description{
Verify that all data are in the same projection
}
\examples{
{
example<-"goes here"
}
}
\seealso{
Other checks: 
\code{\link{.addmainfields2}()},
\code{\link{.checkdatastr2}()},
\code{\link{.checkfields}()},
\code{\link{.status.tracker.and.escaping}()}
}
\author{
Josep M Serra Diaz
}
\concept{checks}
