% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Checks.R
\name{.checkdatastr2}
\alias{.checkdatastr2}
\title{checking main fields}
\usage{
.checkdatastr2(dat, xf = x.field, yf = y.field, verbose = F)
}
\arguments{
\item{dat}{A dataframe containing occurrence data for checking.}

\item{xf}{character. Name of the field where the x coordinate is stored (typically longitude). Default is x.field}

\item{yf}{character. Name of the field where the y coordinate is stored (typically latitude). Default is y.field}
}
\value{
Original dataframe, dat.  Used primarily to generate warning messages.
}
\description{
Verify that all main data fields are correctly populated.
}
\note{
Inspired in addmainfields from biogeo
}
\examples{
\dontrun{
Example<-"goes here"
}

}
\seealso{
Other checks: \code{\link{.addmainfields2}},
  \code{\link{.check.geospatial.data}},
  \code{\link{.checkfields}},
  \code{\link{.status.tracker.and.escaping}}
}
\concept{checks}
