% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Checks.R
\name{.checkfields}
\alias{.checkfields}
\title{Checking main fields}
\usage{
.checkfields(
  dat,
  xf = x.field,
  yf = y.field,
  ef = e.field,
  tf = t.field,
  lf = l.field,
  cf = c.field,
  idf = taxonobservation.id,
  verbose = F
)
}
\arguments{
\item{dat}{A dataframe containing occurrence data for checking.}

\item{xf}{character. Name of the field where the x coordinate is stored (typically longitude). Default is x.field}

\item{yf}{character. Name of the field where the y coordinate is stored (typically latitude). Default is y.field}

\item{ef}{character. Name of the field where the elevation of data collection is stored in the original dataset. Default is e.field.}

\item{tf}{character. Name of the field where the date of data collection is stored in the original dataset. Default is t.field.}

\item{lf}{character. Name of the field where the toponim/location of data collection is stored in the original dataset. Default is l.field.}

\item{cf}{character. Name of the field where the registred country of data collection is stored in the original dataset. Default is c.field.}

\item{idf}{character. Name of the field of the id of the observation}
}
\value{
Original dataframe, dat.  Used primarily to generate warning messages.
}
\description{
Verify that all main data fields are correctly populated.
}
\seealso{
Other checks: 
\code{\link{.addmainfields2}()},
\code{\link{.check.geospatial.data}()},
\code{\link{.checkdatastr2}()},
\code{\link{.status.tracker.and.escaping}()}
}
\concept{checks}
