% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaHulls.R
\name{.checkPolygonsGEOS2}
\alias{.checkPolygonsGEOS2}
\title{Check polygon geometry}
\usage{
.checkPolygonsGEOS2(obj, properly = TRUE, force = TRUE, useSTRtree = FALSE)
}
\arguments{
\item{obj}{an alpha hull object}

\item{properly}{logic.}

\item{force}{logic.}

\item{useSTRtree}{logic.}
}
\value{
a sp polygon object
}
\description{
Check polygon geometry
}
\details{
inspired provided by  maptools package and from P Title in rangeBuilder
}
\seealso{
Alpha hulls are created with \code{\link{ahull}}. \cr
see maptools and RangeBuilder package
}
\author{
Pascal Title (original version), Josep M Serra-Diaz (modifications)
}
