% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Geo.R
\name{.coords2country}
\alias{.coords2country}
\title{Identifies country ISO3 based on locations}
\usage{
.coords2country(points)
}
\arguments{
\item{points}{A dataframe with x and y coordinates}
}
\value{
Factor with ISO3 codes for countries
}
\description{
extracts data from database
}
\examples{
\dontrun{

}

}
\seealso{
Other Geo: \code{\link{.getSRTM}}
}
\concept{Geo}
