% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_OtherFuncs.R
\name{.find_and_transform_dates}
\alias{.find_and_transform_dates}
\title{Identify and harmonize data formats}
\usage{
.find_and_transform_dates(
  data_set,
  cols = "auto",
  formats = NULL,
  n_test = 30,
  ambiguities = "IGNORE",
  verbose = TRUE
)
}
\value{
character
}
\description{
Identify and harmonize data formats
}
\note{

}
\references{

}
\seealso{
dataPreparation::find_and_transform_dates

Other internal: 
\code{\link{.get_os}()},
\code{\link{.gimme.score}()},
\code{\link{.hijack}()},
\code{\link{lazylogic}()},
\code{\link{rm.all}()},
\code{\link{splitSpname}()}
}
\author{

}
\concept{internal}
\keyword{internal}
