% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFuncs.R
\name{.get_os}
\alias{.get_os}
\title{Identify type of Operative system}
\usage{
.get_os()
}
\value{
character
}
\description{
The function is used to report the working operative system as a 3-letter character for 'win', and 'mac', 'unix', and 'unknown OS'
}
\details{
This is implemented whan parallelizing functions
}
\note{
Internal function copied from utils wickham to know which is the os used by the computer
}
\seealso{


Other internal: 
\code{\link{lazylogic}()},
\code{\link{rm.all}()},
\code{\link{splitSpname}()}
}
\author{
Pep Serra-Diaz
}
\concept{internal}
\keyword{internal}
