% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geo.R
\name{.getSRTM}
\alias{.getSRTM}
\title{Download SRTM elevation raster}
\usage{
.getSRTM(xydat, download = T, path = getwd(), verbose = F)
}
\arguments{
\item{xydat}{A dataframe with x and y coordinates}

\item{download}{Default to T. Whether the data should be downloaded}

\item{path}{where the downloads should go. Default to the currendt directory}

\item{verbose}{if you want to print messages of progress or warnings}
}
\value{
raster
}
\description{
Download SRTM elevation raster
}
\details{
Basedd on getData from raster
}
\examples{
\dontrun{
example<-"goes here"
}
}
\seealso{
raster::getData

Other Geo: 
\code{\link{.coords2country}()}
}
\concept{Geo}
