% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_OtherFuncs.R
\name{.gimme.score}
\alias{.gimme.score}
\title{get the scoring outputs from an analysis}
\usage{
.gimme.score(x)
}
\arguments{
\item{x}{data.frame resulting from a series of analysis in occTest}
}
\value{
character
}
\description{
compute the scores for each type of test implemented in occTest function
}
\details{
This is implemented within the occTest function
}
\note{

}
\references{

}
\seealso{


Other internal: 
\code{\link{.find_and_transform_dates}()},
\code{\link{get_os}()},
\code{\link{hijack}()},
\code{\link{lazylogic}()},
\code{\link{rm.all}()},
\code{\link{splitSpname}()}
}
\author{
Pep Serra-Diaz
}
\concept{internal}
\keyword{internal}
