% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occOutliers_modif.R
\name{.iqrOutlier}
\alias{.iqrOutlier}
\title{find records mor than 1.5 times the interquartile range beyond the upper quartile}
\usage{
.iqrOutlier(dists)
}
\arguments{
\item{dists}{a numeric vector}
}
\description{
find records mor than 1.5 times the interquartile range beyond the upper quartile
}
