% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_OtherFuncs.R
\name{.join.spname}
\alias{.join.spname}
\title{Convert to Collated species names}
\usage{
.join.spname(x)
}
\arguments{
\item{x}{character}
}
\value{
character
}
\description{
It collates species names from format "Species genus" to "Species_genus"
}
\examples{
\dontrun{
example<-"goes here"
}
}
\keyword{internal}
