% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-modifiedFunctions.R
\name{.ne_download_occTest}
\alias{.ne_download_occTest}
\title{Download quitely from neearth data (modified from original package)}
\usage{
.ne_download_occTest(
  scale = 110,
  type = "countries",
  category = c("cultural", "physical", "raster"),
  destdir = tempdir(),
  load = TRUE,
  returnclass = c("sp", "sf")
)
}
\arguments{
\item{scale}{Data.frame of species occurrences}

\item{type}{the field in the dataframe containing the x coordinates}

\item{category}{the field in the dataframe containing the y coordinates}

\item{load}{Raster of human influence index}

\item{desetdir}{proj4string argument for dataframe}
}
\value{
spatial object or raster
}
\description{
download from Natural Earth the country checklist
}
\seealso{
\[linkrnaturalearth]{ne_download}
}
\author{
Andy South (southandy@gmail.com) rnaturalearth; Josep M Serra Diaz (modifs)
}
\keyword{internal}
