% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_OtherFuncs.R
\name{.paste3}
\alias{.paste3}
\title{Paste obviating NAs}
\usage{
.paste3(..., sep = ", ")
}
\arguments{
\item{...}{character vector}
}
\value{
character
}
\description{
Removes NAs when pasting character vectors
}
\note{

}
\references{

}
\seealso{

}
\author{
Pep Serra-Diaz
}
\concept{}
\keyword{internal}
