% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occOutliers_modif.R
\name{.presPercentile}
\alias{.presPercentile}
\title{omit outlying pres}
\usage{
.presPercentile(
  xy,
  percent = 95,
  unin = c("m", "km"),
  unout = c("ha", "km2", "m2")
)
}
\arguments{
\item{xy}{data.frame with 2 columns}

\item{percent}{numeric on [0,100]}

\item{unin}{character; input units. option are "m" and "km"}

\item{unout}{character; input units. option are "ha","km2", "m2"}
}
\description{
omit outlying pres
}
