% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Checks.R
\name{.status.tracker.and.escaping}
\alias{.status.tracker.and.escaping}
\title{Workflow status tracker}
\usage{
.status.tracker.and.escaping(
  dataset.to.continue = dat,
  wfo = write.full.output,
  wso = write.simple.output,
  xf = x.field,
  yf = y.field,
  od = output.dir,
  obf = output.base.filename,
  sp = sp.name,
  verbose = F
)
}
\arguments{
\item{dataset.to.continue}{A dataframe containing occurrence data for checking.}

\item{wfo}{write full output}

\item{wso}{write simple output}

\item{xf}{The dataframe field containing the x values (e.g. "longitude")}

\item{yf}{The dataframe field containing the y values (e.g. "latitude")}

\item{od}{The output directory to use}

\item{obf}{Output base filename}

\item{sp}{character. Name of the species}

\item{verbose}{logical. Print messages?}
}
\value{
Original dataframe, dat.  Used primarily to generate warning messages.
}
\description{
Track status and write useful output
}
\examples{
\dontrun{
Example<-"goes here"
}
}
\seealso{
Other checks: 
\code{\link{.addmainfields2}()},
\code{\link{.check.geospatial.data}()},
\code{\link{.checkdatastr2}()},
\code{\link{.checkfields}()}
}
\concept{checks}
