% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_OtherFuncs.R
\name{.subsetlist.nonNULL}
\alias{.subsetlist.nonNULL}
\title{Subset to non-NULL lists}
\usage{
.subsetlist.nonNULL(x)
}
\arguments{
\item{x}{List}
}
\value{
list
}
\description{
Subset a list removing NULL entries
}
\examples{
\dontrun{
example<-"goes here"
}
}
\keyword{internal}
