% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occOutliers_modif.R
\name{.vgrep}
\alias{.vgrep}
\title{Vectorized version of grep}
\usage{
.vgrep(pattern, x)
}
\arguments{
\item{pattern}{character string containing a regular expression (or character string for ‘fixed = TRUE’) to be matched in the given character vector.  Coerced by ‘as.character’ to a character string if possible.  If a character vector of length 2 or more is supplied, the first element is used with  a warning.}

\item{x}{a character vector where matches are sought, or an object which can be coerced by ‘as.character’ to a character vector.}
}
\description{
vectorized version of grep
}
