% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{filterMissing}
\alias{filterMissing}
\title{Check for missing coordinates}
\usage{
filterMissing(df, xf = x.field, yf = y.field, verbose = F)
}
\arguments{
\item{df}{dat.frame of species occurrences}

\item{xf}{character. The field in the dataframe containing the x cordinates}

\item{yf}{character. The field in the dataframe containing the y cordinates}
}
\value{
List with two dataframes: stay = coordinates missing and continue = occurrence that you can retain for further analysis
}
\description{
checks for missing coordinates in the occurrence dataframe
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{internal}
