% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occOutliers_modif.R
\name{findEnvOutliers}
\alias{findEnvOutliers}
\title{Find outlying occurrence data in environmental space}
\usage{
findEnvOutliers(
  pres,
  pvalSet = 1e-05,
  method = "grubbs",
  checkPairs = TRUE,
  kRosner = NULL
)
}
\arguments{
\item{pres}{a `SpatialPoints` or `SpatialPointsDataFrame` object describing the locations of species records. A `SpatialPointsDataFrame` containing the values of environmental variables to be used must be supplied if `envOutliers=TRUE`}

\item{pvalSet}{user-specified p-value for assessing the significance of Grubbs test statistic.}

\item{method}{character; options are 'iqr', 'grubbs', 'dixon', 'rosner'}

\item{checkPairs}{logical; check for a single pair of outliers using the Grubbs test. This can only be performed for sample sizes <30. Only a single test is used because repeating it tends to throw out more points than seem reasonable, by eye. The value has no effect unless `method='grubbs'`.}

\item{kRosner}{integer between 1 and 10. Determines the number of outliers suspected with a Rosner test. The value has no effect unless `method='rosner'`.}
}
\value{
Returns the indices of environmental outliers
}
\description{
Environmental outliers
}
\details{
See Examples.
}
\examples{
myPres=read.csv(system.file('extdata/SpeciesCSVs/Camissonia_tanacetifolia.csv',
                            package='occOutliers'))
myPres=myPres[complete.cases(myPres),]
sp::coordinates(myPres)=c(1,2)
myEnv=terra::rast(system.file('extdata/AllEnv.tif',package='occOutliers'))
names(myEnv)=read.table(system.file('extdata/layerNames.csv',package='occOutliers'))[,1]
myPresDF=sf::st_as_sf(myPres,data.frame(terra::extract(myEnv,myPres)))
presOut=findEnvOutliers(pres=myPresDF,pvalSet=1e-5)
}
\author{
Cory Merow <cory.merow@gmail.com>
}
