% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{geoEnvAccuracy}
\alias{geoEnvAccuracy}
\title{Coordinate accuracy}
\usage{
geoEnvAccuracy(
  df,
  xf,
  yf,
  af,
  dsf = NULL,
  ef,
  method = "all",
  r.env,
  accept.threshold.cell = 0.5,
  accept.threshold.env = 0.5,
  bearing.classes = 10,
  distance.classes = 5,
  env.quantiles = c(0.3, 0.7),
  elev.threshold = 100,
  raster.elevation = NULL,
  verbose = FALSE,
  do = TRUE,
  doParallel = FALSE,
  mc.cores = 2
)
}
\arguments{
\item{df}{\emph{data.frame}. Dataset of species occurrences with coordinates}

\item{xf}{\emph{character}. column name in df containing the x coordinates}

\item{yf}{\emph{character}. column name in df containing the y coordinates}

\item{af}{\emph{character}. column name in df containing the coordinate uncertainty value (in the same)}

\item{dsf}{\emph{character}. column name in df containing the dataset to which the record belongs to (e.g. Forest Inventory of Spain)}

\item{ef}{\emph{character}. column name in df containing the registered elevation for the record.}

\item{method}{\emph{character}. Vector of methods to be used. See details. Default 'all'}

\item{r.env}{\emph{SpatRaster}. Raster with environmental data, typically temperature.}

\item{accept.threshold.cell}{\emph{numeric}. Acceptance threshold for how much percentage of the Area of uncertainty in the cell we want to accept. Default to 0.5}

\item{accept.threshold.env}{\emph{numeric}. Acceptance threshold for how much percentage difference in an environmental variable you are willing to accept given a geographic accuracy. Default to 0.5}

\item{bearing.classes}{\emph{integer}. Directions to analyze given accuracy. Default to 10.}

\item{distance.classes}{\emph{integer}. Number of distance classes to analyze in terms of geographic accuracy. Default to 5.}

\item{env.quantiles}{\emph{numeric}. Max and min quantiles to identify environmental inaccuracies.  Default to c(0.3,0.7)}

\item{elev.threshold}{\emph{numeric}. Maximum difference accepted between target cell and cells within the geographic inaccuracy. Default to 100}

\item{raster.elevation}{\emph{SpatRaster}  Raster of a digital elevation model.}

\item{verbose}{\emph{logical}. Print messages? Default FALSE}

\item{do}{\emph{logical}. Should tests be performed? Default TRUE}

\item{doParallel}{\emph{logical}. Should computation use parallel functions? Default FALSE}

\item{mc.cores}{\emph{numeric}. How many cores to use? (used when doParallel = TRUE). Default 2}
}
\value{
a \emph{data.frame} with the results of the tests
}
\description{
Detect records with low accuracy affecting environmental values
}
\details{
Geoenvironmental accuracy function will implement differnt methods to assess occurrence accuracy in environmnental and geographic space.\cr
Current implmented methods are:
'lattice' : tests for lattice arrangement in occurrence datasets. Borrowed from \link[CoordinateCleaner]{cd_round} . \cr
'elevDiff' : assess the elevation difference between a given raster (or automatically downloaded fro SRTM), and the elevation recorded. If differences >elev.threshold then the record is considered as a low accuracy threshold\cr
'noDate' : assess whether there is a date or timestamp information in the record. \cr 
'noDateFormatKnown' : assess whether the information in the timestamp agrees with different formatting of Dates. \cr 
'outDateRange' : (not implemented) assess whether the record is within a user specified time frame. \cr 
'percDiffCell' : assess whether the record may be falling in a different raster cell given an information of coordinate accuracy. \cr 
'envDeviation' : assess whether the climate in a given record may be outside of the interval 30th-70th (default values) for a given variable due to coordinate uncertainty. \cr
}
\examples{
\dontrun{

sp_file <- system.file('ext/exampleOccData.csv',package='occTest')
occ <- read.table(file = sp_file,header = TRUE,sep = ',',as.is = TRUE,nrows = 50)
env <- terra::rast(system.file('ext/AllEnv.tif',package='occTest'))
dem <- terra::rast(system.file('ext/DEM.tif',package='occTest'))
geoEnvAccuracy(df=occ,xf = 'MAPX',yf = 'MAPY',
                af = c('UNCERTAINTY_X_M','UNCERTAINTY_Y_M'),
                ef = 'ELEVATION',
                r.env = env[[1]],
                raster.elevation = dem)
#see more examples in vignetteXtra-occTest
#devtools::install_github ('vignetteXtra-occTest')
}
}
\seealso{
\link[CoordinateCleaner]{cd_round}

Other analysis: 
\code{\link{.nearestcell3}()},
\code{\link{biogeo_outliers}()},
\code{\link{centroidDetection}()},
\code{\link{countryStatusRangeAnalysis}()},
\code{\link{duplicatesexcludeAnalysis}()},
\code{\link{humanDetection}()},
\code{\link{timeAccuracy}()}
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr), A Zizka (CoordinateCleaner package)
}
\concept{analysis}
\keyword{Analysis}
