% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{geoOutliers}
\alias{geoOutliers}
\title{ALPHA HULL OUTLIERS}
\usage{
geoOutliers(
  df = dat,
  xf = x.field,
  yf = y.field,
  .alpha.parameter = alpha.parameter,
  .distance.parameter = 1000,
  .medianDeviation.parameter = 5,
  .samplingIntensThreshold.parameter = 0.1,
  .projString = points. | envOutliers | "bxp" | proj4string,
  method = "all",
  do = T,
  verbose = F
)
}
\arguments{
\item{df}{Data.frame of species occurrences}

\item{xf}{the field in the dataframe containing the x cordinates}

\item{yf}{the field in the dataframe containing the y cordinates}

\item{.alpha.parameter}{parameter setting for alphahull}
}
\value{
list
}
\description{
ALPHA HULL OUTLIERS
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{internal}
