% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_occTest_settings}
\alias{get_occTest_settings}
\title{Get occTest Settings}
\usage{
get_occTest_settings(x)
}
\arguments{
\item{x}{An occTest or occFilter object returned by  {\link[=occTest]{occTest}} or  {\link[=occFilter]{occFilter}}}
}
\value{
list of lists with all different parameters to use in  {\link[=occTest]{occTest}}
}
\description{
Get the settings used to create a  occTest or occFilter object
}
\examples{
\dontrun{
example<-"goes here"
}
}
\seealso{
{\link[=occTest]{occTest}}; {\link[=occFilter]{occFilter}}
}
\author{
J Borderieux (jeremy.borderieux@agroparistech.fr)
}
\keyword{occTest}
