% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_OtherFuncs.R
\name{get_os}
\alias{get_os}
\title{Checks the Operative System}
\usage{
get_os()
}
\value{
a character function.
}
\description{
Checks the Operative System
}
\note{
#TInternal function copied from utiles wickham to know which is the os used by the computer
}
\examples{
\dontrun{
example<-"goes here"
}
}
