% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_OtherFuncs.R
\name{hijack}
\alias{hijack}
\title{Hijacking functions to change default parameters}
\usage{
hijack(FUN, ...)
}
\value{
a function
}
\description{
Hijacking functions to change default parameters
}
\note{
#Internal function that I got from stack overflow MrFlick
}
\examples{
\dontrun{
.data.frame <- hijack(data.frame, stringsAsFactors = FALSE)
dat <- .data.frame(x1 = 1:3, x2 = c("a", "b", "c"))
str(dat)  # yay! strings are character
}


}
