% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{hyperHumanDetection}
\alias{hyperHumanDetection}
\title{HYPER HUMAN ENVIRONMENT FUNCTION}
\usage{
hyperHumanDetection(df = dat, xf = x.field, yf = y.field,
  method = "all", .points.proj4string = points.proj4string,
  ras.hii = ras.hii, .th.human.influence = th.human.influence,
  do = T, verbose = F)
}
\arguments{
\item{df}{Data.frame of species occurrences}

\item{xf}{the field in the dataframe containing the x cordinates}

\item{yf}{the field in the dataframe containing the y cordinates}

\item{.points.proj4string}{proj4string argument for dataframe}

\item{ras.hii}{Raster of human influence index}

\item{.th.human.influence}{threshold of human influence index}
}
\value{
list
}
\description{
Detect occurrences in heavily human-impacted environments
}
\examples{
\dontrun{
example<-"goes here"
}
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{internal}
