% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{institutionLocality}
\alias{institutionLocality}
\title{POTENTIAL BOTANIC GARDEN LOCALITY FROM NAME}
\usage{
institutionLocality(df = dat, xf = x.field, yf = y.field,
  .points.proj4string = points.proj4string, lf = l.field,
  method = "all", do = T, verbose = F)
}
\arguments{
\item{df}{Data.frame of species occurrences}

\item{lf}{The locality field in the dataframe, df.}
}
\value{
list
}
\description{
Detect occurrences potentially in botanical gardens via locality name
}
\examples{
\dontrun{
example<-"goes here"
}
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{internal}
