% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landFilter.R.R
\name{landSeaFilter}
\alias{landSeaFilter}
\title{Filter occurrence by habitat (terrestrial/non-Terrestrial)}
\usage{
landSeaFilter(df, xf, yf, habType = NULL, verbose = TRUE, habPol = NULL)
}
\arguments{
\item{df}{\emph{data.frame} of species occurrences}

\item{xf}{\emph{character}. The field in df containing the x coordinates}

\item{yf}{\emph{character}. The field in df containing the y coordinates}

\item{habType}{\emph{character}. Define the species habitat. Only "terrestrial" and "sea" implemented.}

\item{verbose}{\emph{logical} Print messages? Default TRUE}

\item{habPol}{\emph{polygon sf}. Multipolygon object indicating land masses}
}
\value{
\emph{list} of filtered and non-filtered occurrrence recprds by land use type
}
\description{
Filter the occurrence records according to whether they should be in land masses or not
}
\details{
The user can define the habitat type and provide a an input landmass polygon
}
\examples{
\dontrun{
xyDF <- data.frame (x=c(0,42),y=c(0,1),Reason=NA)
landSeaFilter(df=xyDF,xf='x',yf='y')
}
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{filter}
