% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{landUseSelect}
\alias{landUseSelect}
\title{Selection of records within a land use selectino}
\usage{
landUseSelect(
  df = dat,
  xf = x.field,
  yf = y.field,
  method = "in",
  .points.proj4string = points.proj4string,
  ras.landUse = ras.landUse,
  .landUseCodes = landUseCodes,
  do = T,
  verbose = F,
  output.dir = output.dir
)
}
\value{
data.frame
}
\description{
Selection of records within a land use selectino
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{internal}
