% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_OtherFuncs.R
\name{lazylogic}
\alias{lazylogic}
\title{a lazy logic omitting NAs}
\usage{
lazylogic(e)
}
\arguments{
\item{e}{expression}
}
\value{
logic
}
\description{
evaluates a conditional expression where if any element is NA the output is F instead of NA
}
\examples{
\dontrun{
 k = NA
 k == 1
 lazylogic (k==1)
}
}
\seealso{
Other internal: 
\code{\link{.find_and_transform_dates}()},
\code{\link{.gimme.score}()},
\code{\link{get_os}()},
\code{\link{hijack}()},
\code{\link{rm.all}()},
\code{\link{splitSpname}()}
}
\concept{internal}
