% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{mcp_test}
\alias{mcp_test}
\title{Find outlying occurrence data in geographical space from convex hulls}
\usage{
mcp_test(data, xcoord, ycoord, percentage = 95)
}
\arguments{
\item{data}{\emph{data.frame}  with coordinate information}

\item{xcoord}{\emph{character}. Column with X coordinate information in data}

\item{ycoord}{\emph{character}. Column with Y coordinate information in data}

\item{percentage}{\emph{numeric}. Percentage of data used to build the minimum convex polygon}
}
\value{
Returns a \emph{logical} vectors detecting outliers
}
\description{
Computes minimum convex hulls and identifies outlying points
}
\details{
only works for WGS84 lat long
}
\examples{
myPres<-read.csv(system.file('./ext/exampleOccData.csv',package='occTest'))|>
  dplyr::select(c('MAPX','MAPY'))
mcp_test(data = myPres,xcoord = 'MAPX',ycoord = 'MAPY')
}
\seealso{
\link[occTest]{geoOutliers}
}
\author{
Coline CF Boonmann <colineboonman@bio.au.dk>
}
\keyword{internal}
