% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_modifiedFunctions.R
\name{ne_download_occTest}
\alias{ne_download_occTest}
\title{Download (QUIETELY!) from neearth data (modified from original package)}
\usage{
ne_download_occTest(
  scale = 110,
  type = "countries",
  category = c("cultural", "physical", "raster"),
  destdir = tempdir(),
  load = TRUE,
  returnclass = c("sp", "sf")
)
}
\arguments{
\item{scale}{Data.frame of species occurrences}

\item{type}{the field in the dataframe containing the x cordinates}

\item{category}{the field in the dataframe containing the y cordinates}

\item{load}{Raster of human influence index}

\item{desetdir}{proj4string argument for dataframe}
}
\value{
spatial object or raster
}
\description{
download from Natural Earth
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
author of the package neearth
}
\concept{}
\keyword{internal}
