% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{nearestcell3}
\alias{nearestcell3}
\title{SeaLand reassignement}
\usage{
nearestcell3(dat, rst, xf = x.field, yf = y.field, verbose = F)
}
\arguments{
\item{dat}{Data.frame of species occurrences and a field named taxonobservationID}

\item{rst}{An optional raster grid}

\item{xf}{the field in the dataframe containing the x cordinates}

\item{yf}{the field in the dataframe containing the y cordinates}
}
\value{
list
}
\description{
SeaLand reassignement
}
\details{
(function inspiered in nearestcell in biogeo bu modified)
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{internal}
