% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_wrappers.R
\name{occSimpleClassification}
\alias{occSimpleClassification}
\title{Runs tests and validates data}
\usage{
occSimpleClassification(
  spOcc,
  env,
  speciesName = "My species",
  x = "x",
  y = "y",
  date = NULL,
  isoCountry = NULL,
  classification = NULL
)
}
\arguments{
\item{spOcc}{data.frame. Object with the coordinate data.}

\item{env}{raster or rasterStack. Environmental data (e.g. typically climatic).}

\item{speciesName}{character. Name of the species.}

\item{x}{name of the field with the coordinate x. Default 'x'}

\item{y}{name of the field with the coordinate y. Default 'y'}

\item{isoCountry}{name of the field with the values of the reported country of the record. Default NULL}

\item{classification}{character. Indicates the thresholds phylosofphy applied to classify errors in occurrence data. Possible values 'strict','relaxed','custom'}

\item{data}{name of the field with the values of the reported timestamp of the record. Default NULL}
}
\value{
a list of two. First element is a dataframe with profiled occurrence records with their associated profiled labels. Second element is a dataframe with all outputs of the analysis implemented.
}
\description{
Runs tests and validates data
}
\note{
#There are several parameters in the function. The majority of them can be adjusted, but we also provide default values. We recommend those default values if the user is to use the geospatial data included in the package.
}
\examples{
\dontrun{
example<-"goes here"
}
}
