% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_workflow.R
\name{occurrenceClassify}
\alias{occurrenceClassify}
\title{Validates and clean occurrence data}
\usage{
occurrenceClassify(
  sp.name,
  sp.table,
  r.env,
  tableSettings = NULL,
  analysisSettings = NULL,
  gradingSettings = NULL,
  writeoutSettings = NULL,
  r.dem = NULL,
  ntv.ctry = NULL,
  inv.ctry = NULL,
  resolveAlienCtry = F,
  resolveNativeCtry = F,
  interactiveMode = F,
  verbose = F
)
}
\arguments{
\item{sp.name}{character. Name of the species.}

\item{sp.table}{data.frame. Object with the coordinate data.}

\item{r.env}{raster or rasterStack. Environmental data (e.g. typically climatic).}

\item{tableSettings}{list. Elements corresponding to different settings of the input occurrence table.}

\item{analysisSettings}{list. Elements corresponding to different settings of the analysis functions .}

\item{gradingSettings}{list. Elements corresponding to different settings of the analysis functions .}

\item{writeoutSettings}{list. Elements corresponding to different settings of the analysis functions .}

\item{r.dem}{raster. Elevation data (in meters).}
}
\value{
a list of two. First element is a dataframe with profiled occurrence records with their associated profiled labels. Second element is a dataframe with all outputs of the analysis implemented.
}
\description{
occurrence.profiler returns occurrence points withou the recommended metadata and attribution information.
}
\note{
#There are several parameters in the function. The majority of them can be adjusted, but we also provide default values. We recommend those default values if the user is to use the geospatial data included in the package.
}
\examples{
\dontrun{
example<-"goes here"
}
}
