% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_workflow.R
\name{occurrenceProfile}
\alias{occurrenceProfile}
\title{Validates and clean occurrence data}
\usage{
occurrenceProfile(
  output.dir = NULL,
  sp.table,
  sp.name,
  taxonobservation.id = NULL,
  x.field = "x",
  y.field = "y",
  t.field = NULL,
  l.field = NULL,
  c.field = NULL,
  e.field = NULL,
  a.field = NULL,
  ds.field = NULL,
  r.env,
  r.dem = NULL,
  countries.shapefile = NULL,
  countryfield.shapefile = "ISO",
  ntv.ctry = NULL,
  inv.ctry = NULL,
  ras.hii = NULL,
  coordinate.decimal.precision = 4,
 
    points.proj4string = sp:::CRS("+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"),
  doRangeAnalysis = T,
  excludeUnknownRanges = F,
  excludeNotmatchCountry = F,
  doCountryRecordAnalysis = T,
  doCentroidDetection = T,
  methodCentroidDetection = "all",
  doHyperHumanDetection = T,
  methodHyperHumanDetection = "all",
  doInstitutionLocality = T,
  methodInstitutionLocality = "all",
  doGeoOutliers = T,
  methodGeoOutliers = "all",
  doEnvOutliers = T,
  methodEnvOutliers = "all",
  methodGeoEnvAccuracy = "all",
  do.geoEnvAccuracy = T,
  alpha.parameter = 2,
  th.human.influence = 45,
  th.perc.outenv = 0.2,
  elev.quality.threshold = 100,
  grading.test.type = "majority",
  qualifiers = T,
  qualifier.label.scoping = c("A", "B", "C", "D", "E"),
  write = F,
  write.simple.output = F,
  write.full.output = F,
  output.base.filename = "QAQC",
  verbose = F
)
}
\arguments{
\item{output.dir}{Character. The folder where the output will be dumped.}

\item{sp.table}{data.frame. Object with the coordinate data.}

\item{sp.name}{character. Name of the species.}

\item{taxonobservation.id}{characther. field with the taxon observation identifier (default = NULL).}

\item{x.field}{character. Name of the field where the x coordinate is stored (typically longitude).}

\item{y.field}{character. Name of the field where the y coordinate is stored (typically latitude).}

\item{t.field}{character. Name of the field where the date of data collection is stored in the original dataset. Default NULL.}

\item{l.field}{character. Name of the field where the toponim/location of data collection is stored in the original dataset. Default NULL.}

\item{c.field}{character. Name of the field where the registred country of data collection is stored in the original dataset. Default NULL.}

\item{e.field}{character. Name of the field where the elevation of data collection is stored in the original dataset. Default NULL.}

\item{r.env}{raster or rasterStack. Environmental data (e.g. typically climatic).}

\item{r.dem}{raster. Elevation data (in meters).}

\item{countries.shapefile}{SpatialPolygonsDataframe. Country borders with associated data on country name, etc. etc.}

\item{countryfield.shapefile}{character. Name of the field in **_countries.shapefile_** where the 3 letter ISO is stored. Default is "ISO".}

\item{ntv.ctry}{character. Three letter code of the country/countries were the species is native. If set to NULL then all occurrences will be considered as in the native range.}

\item{inv.ctry}{character. Three letter code of the country/countries were the species is invasive If set to NULL then the species will never considered invasive anywhere.}

\item{ras.hii}{raster. Raster with associated values informing on the degree of anthropogenic influence in a place.}

\item{coordinate.decimal.precision}{numeric. Decimal precision of the coordinates. Default is 4.}

\item{points.proj4string}{CRS object. Projection of the orginal presence coordinates. Set toCRS("+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"). This option can't be changed at the moment.}

\item{alpha.parameter}{numeric. Parameter for determining the alpha.hull for geographic outlier detection. Default is 2.}

\item{th.human.influence}{numeric. Maximum level of anthropogenic influence for anthropogenic issues detection. Default is 45.}

\item{th.perc.outenv}{numeric. Ratio of variables considered outliers for evnironmental spaces issues detection. Default is 0.2.}

\item{elev.quality.threshold}{numeric. Maximum difference between recorded and expected elevation allowed for elevation quality assessement. Default is 100.}

\item{qualifiers}{logical. Compute qualifiers. Deafault is T.}

\item{qualifier.label.scoping}{character. Quality grades over which to compute the qualifiers. Default is "A".}

\item{write.simple.output}{logical. Write output table with basic information of the grading. Deafault is F.}

\item{write.full.output}{logical. Write output table with full information of the grading. Deafault is F.}

\item{output.base.filename}{character. Base filenames for outputs. Meaningfull if write outputs are set to T. Default is "QAQC"}
}
\value{
a list of two. First element is a dataframe with profiled occurrence records with their associated profiled labels. Second element is a dataframe with all outputs of the analysis implemented.
}
\description{
occurrence.profiler returns occurrence points withou the recommended metadata and attribution information.
}
\note{
#There are several parameters in the function. The majority of them can be adjusted, but we also provide default values. We recommend those default values if the user is to use the geospatial data included in the package.
}
\examples{
\dontrun{
example<-"goes here"
}
}
