% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.occTest}
\alias{plot.occTest}
\title{Display the filtering process}
\usage{
\method{plot}{occTest}(x, occFilter_list = NULL, show_plot = F, ...)
}
\arguments{
\item{x}{An  occTest object returned by  {\link[=occTest]{occTest}}, i.e. the unfiltered data.frame}

\item{occFilter_list}{Optional, an occFilter object; a list returned by  {\link[=occFilter]{occFilter}}, the result of the filtering of \code{x}}

\item{show_plot}{Logical, should the plots be plotted ?}

\item{...}{not used}
}
\value{
list of ggplots objects, of varying length, depending on whether the filtering was done by testBlock or testType
}
\description{
Display the filtering process
}
\details{
If \code{occFilter_list} is provided, display how the occurences passed the different tests, otherwise only plot the coordinates filtering step
}
\examples{
\dontrun{
example<-"goes here"
}
}
\seealso{
{\link[=occFilter]{occFilter}}  , {\link[=occTest]{occTest}}  , the {\link[=ggplot2]{ggplot2}} package
}
\author{
J Borderieux (jeremy.borderieux@agroparistech.fr)
}
\keyword{filter}
\keyword{plot}
