% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_plot.R
\name{plot_occFilter}
\alias{plot_occFilter}
\title{Display the filtering process}
\usage{
plot_occFilter(occfilter_list, occTest_df, show_plot = T)
}
\arguments{
\item{df}{data.frame outputed from occurrenceClassify}

\item{by}{character. Applying thresholds to either  blocks of test ('testBlock') or single test types (testTypes )}

\item{threshold}{character. Phylosophy for filtering based on threshold. Option are majority, relaxed, stringent}

\item{customThresholds}{data.frame. Specify for each dimension of flags the specific threshold to use. It overides the parameters in thresholds. We recomend building that table based on the functio buildCustomThresholds.}
}
\value{
list of 2 data.frames
}
\description{
Display the filtering process
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{filter}
\keyword{plot}
