% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Unsorted_Functions.R
\name{presPercentile}
\alias{presPercentile}
\title{Identify percentile of presences}
\usage{
presPercentile(
  xy,
  percent = 95,
  unin = c("m", "km"),
  unout = c("ha", "km2", "m2")
)
}
\arguments{
\item{x}{raster* object}

\item{...}{additional functions to be passed to \code{\link[raster]{writeRaster}}}
}
\description{
Divide raster by the sum of all cells.
}
