% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_plot.R
\name{printOccTest}
\alias{printOccTest}
\title{Display the filtering process}
\usage{
printOccTest(x, occFilter_list = NULL, show_plot = F)
}
\arguments{
\item{x}{An  occTest object returned by  {\link[=occTest]{occTest}}, i.e. the unfiltered data.frame}

\item{occFilter_list}{Optional, an occFilter object; a list returned by  {\link[=occFilter]{occFilter}}, the result of the filtering of \code{x}}

\item{show_plot}{Logical, should the plots be plotted ?}
}
\value{
list of ggplots objects, of varying length, depending on whether the filtering was done by testBlock or testType
}
\description{
#' @title Display the filtering process
#' @descriptions Display the filtering process 
#' @details If \code{occFilter_list} is provided, display how the occurences passed the different tests, otherwise only plot the coordinates filtering step
#' @return list of ggplots objects, of varying length, depending on whether the filtering was done by testBlock or testType
#' @keywords plot filter
#' @author J Borderieux (jeremy.borderieux@agroparistech.fr)
#' @param x An  occTest object returned by  {\link[=occTest]{occTest}}, i.e. the unfiltered data.frame 
#' @param occFilter_list Optional, an occFilter object; a list returned by  {\link[=occFilter]{occFilter}}, the result of the filtering of \code{x}
#' @param show_plot  Logical, should the plots be plotted ?
#' @note
#' @seealso  {\link[=occFilter]{occFilter}}  , {\link[=occTest]{occTest}}  , the {\link[=ggplot2]{ggplot2}} package
#' @references
#' @aliases
#' @family
#' @method plot occTest
#' @examples \dontrun{
#' example<-"goes here"
#' }
#' @export
}
\details{

}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{
{\link[=occFilter]{occFilter}}  , {\link[=occTest]{occTest}}  , the {\link[=ggplot2]{ggplot2}} package
}
\author{
J Borderieux (jeremy.borderieux@agroparistech.fr)
}
\concept{}
\keyword{filter}
\keyword{plot}
