% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reTest.R
\name{reTest}
\alias{reTest}
\title{Updates occTest with new user-defined tests}
\usage{
reTest(occTest_result, my_new_test)
}
\arguments{
\item{occTest_result}{\emph{data.frame} or tibble. Output from an occTest project}

\item{my_new_test}{\emph{data.frame} or \emph{tibble}. New tests following the syntax of occTest (see Details)}
}
\value{
\emph{data.frame with} the new user-defined tests incorporated in the table and the testType_scores columns updated.
}
\description{
Reorganize an output from occTest columns to incorporate user defined tests.
}
\details{
my_test needs to follow the syntax of occTest testType_testMethod_test, 
testType_testMethod_value, testType_testMethod_comment 
the parameter my_test needs to have at least a logical column with 
the convention [testType]_[testMethod]_[test]
where testTypes are the different tests types implemented in occTest
}
\examples{
\donttest{
### THIS IS A CUT DOWN  EXAMPLE 
### visit vignetteXtra-occTest for more info
out <- readRDS (system.file('ext/output_occTest.rds',package = 'occTest'))
myNewTest <- data.frame (humanDetection_myInventedTest_test=sample (c(TRUE,FALSE),
                                                                    replace = TRUE,
                                                                    size = nrow(out)))
out_test_new <- reTest(occTest_result = out, my_new_test = myNewTest)
out_test_new
}
}
