% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_OtherFuncs.R
\name{rm.all}
\alias{rm.all}
\title{Remove all objects}
\usage{
rm.all()
}
\value{
a clear environment
}
\description{
removes all objects in the environment
}
\examples{
\dontrun{
}
}
\seealso{
Other internal: 
\code{\link{.find_and_transform_dates}()},
\code{\link{.gimme.score}()},
\code{\link{get_os}()},
\code{\link{hijack}()},
\code{\link{lazylogic}()},
\code{\link{splitSpname}()}
}
\concept{internal}
