% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{set_testTypes}
\alias{set_testTypes}
\title{Set the tests to run}
\usage{
set_testTypes(
  countryStatusRange = TRUE,
  centroidDetection = TRUE,
  humanDetection = TRUE,
  landUse = TRUE,
  institutionLocality = TRUE,
  geoOutliers = TRUE,
  envOutliers = TRUE,
  geoenvLowAccuracy = TRUE,
  timeAccuracy = TRUE
)
}
\arguments{
\item{countryStatusRange}{logical. Should this test type be performed?}

\item{centroidDetection}{logical. Should this test type be performed?}

\item{humanDetection}{logical. Should this test type be performed?}

\item{institutionLocality}{logical. Should this test type be performed?}

\item{geoOutliers}{logical. Should this test type be performed?}

\item{envOutliers}{logical. Should this test type be performed?}

\item{geoenvLowAccuracy}{logical. Should this test type be performed?}

\item{timeAccuracy}{logical. Should this test type be perfomed?}

\item{landUseType}{logical. Should this test type be performed?}
}
\value{
a \emph{list} with the parameter settings to run \link[occTest]{occTest}
}
\description{
function used to select which types of tests you want in occTest workflow (analysisSet)
}
\details{
See occTest::showTests for further information on tests used in the packages
}
\examples{
\dontrun{
#now we do not want to perform centroid geo-environmental accuracy type of tests
mySettings_analysis <- set_testTypes(geoenvLowAccuracy=FALSE)
}
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{user}
