% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_defaultSettings.R
\name{setTableNames}
\alias{setTableNames}
\title{set table names internally}
\usage{
setTableNames(
  x.field = NULL,
  y.field = NULL,
  t.field = NULL,
  l.field = NULL,
  c.field = NULL,
  e.field = NULL,
  a.field = NULL,
  ds.field = NULL,
  taxonobservation.id = NULL
)
}
\value{
list
}
\description{
set table names internally
}
\examples{
\dontrun{
example<-"goes here"
}
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{user}
