% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_defaultSettings.R
\name{setTests}
\alias{setTests}
\title{set the important parameters}
\usage{
setTests(
  countryStatusRange = T,
  centroidDetection = T,
  humanDetection = T,
  landUseType = T,
  institutionLocality = T,
  geoOutliers = T,
  envOutliers = T,
  geoenvLowAccuracy = T
)
}
\value{
list
}
\description{
set the important parameters
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{user}
