% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_analysisSettings}
\alias{show_analysisSettings}
\title{Print naming conventions and definitions for analysisSettings list}
\usage{
show_analysisSettings()
}
\value{
prints a data.frame with the analysis settings
}
\description{
prints a table with the the conventions used for analysisSettings list
}
\details{
The function prints a guide to the list naming conventions used by occTest in their default parameters. 
         These defaults can be changed manually by the use. To follow the same structure we recommend loading defaultSettings() and then modify $analysisSettings based on that structure 
         It does not require input parameters
}
\examples{
show_analysisSettings ()
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{user}
