% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_tests}
\alias{show_tests}
\title{Show implemented tests and types of tests}
\usage{
show_tests()
}
\value{
prints a dataframe
}
\description{
prints a table with the column names
}
\details{
The function prints a guide to column naming conventions used by occTest in their default parameters. These defaults can be changed via set_tableNames, but the user may also decide to format their input table according to these naming conventions.
}
\examples{
show_tests()
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{user}
