% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_writeoutSettings}
\alias{show_writeoutSettings}
\title{Print naming conventions and definitions for writoutSettings list}
\usage{
show_writeoutSettings()
}
\value{
prints a data.frame
}
\description{
prints a table with the the conventions used for writoutSettings list
}
\details{
The function prints a guide to the list naming conventions used by occTest in their default parameters. 
         These defaults can be changed via set_writeout. 
         It does not require input parameters
}
\examples{
show_tableNames ()
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{user}
