% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultSettings.R
\name{showTableNames}
\alias{showTableNames}
\title{Print naming conventions in occTest}
\usage{
showTableNames()
}
\value{
prints a data.frame
}
\description{
prints a table with the the conventions used for column names
}
\details{
The function prints a guide to column naming conventions used by occTest in their default parameters. 
         These defaults can be changed via setTableNames, but the user may also decide to format their input table according to these naming conventions. 
         It does not require input parameters
}
\examples{
\dontrun{
example<-"goes here"
}
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{user}
