% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultSettings.R
\name{showTests}
\alias{showTests}
\title{Show implemented tests and types of tests}
\usage{
showTests()
}
\value{
prints a dataframe
}
\description{
prints a table with the column names
}
\details{
The function prints a guide to column naming conventions used by occTest in their default parameters. These defaults can be changed via setTableNames, but the user may also decide to format their input table according to these naming conventions.
}
\examples{
\dontrun{
example<-"goes here"
}
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{user}
