% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_OtherFuncs.R
\name{splitSpname}
\alias{splitSpname}
\title{separate species name}
\usage{
splitSpname(x)
}
\arguments{
\item{x}{species name join as a string separated by "_"}
}
\value{
a species name where genus and species are separated by a space
}
\description{
separate species name
}
\note{
#Internal function
}
\examples{
\dontrun{
}


}
