% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{centroid_assessment}
\alias{centroid_assessment}
\title{Centroid assessment using BIEN methods}
\usage{
centroid_assessment(occurrences, centroid_data)
}
\value{
list
}
\description{
Centroid assessment using BIEN methods
}
\examples{
\dontrun{
example<-"goes here"
}
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{internal}
