% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{centroid_assessment}
\alias{centroid_assessment}
\title{Centroid assessment using BIEN methods}
\usage{
centroid_assessment(occurrences, centroid_data)
}
\arguments{
\item{occurrences}{data.frame with occurrence data}

\item{centroid_data}{data.frame with centroid coordinates for different administrative entities.}
}
\value{
data.frame
}
\description{
Centroid assessment using BIEN methods
}
\author{
Brian Maitner
}
\keyword{internal}
