% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_assessment_BIEN.R
\name{centroid_assessment_BIEN}
\alias{centroid_assessment_BIEN}
\title{Identify occurrences via BIEN proximity analysis}
\usage{
centroid_assessment_BIEN(
  occurrences,
  centroid_data,
  relative_distance_threshold = 0.001
)
}
\arguments{
\item{occurrences}{A occurrence sf data frame containing the correct fields. See example for more details.}

\item{centroid_data}{Centroid sf data.frame from BIEN database}

\item{relative_distance_threshold}{Threshold below which coordinates are assumed to be centroids. Default value is...}
}
\value{
A data.frame containing information about centroid inference.  
is_centroid: Logical.  1 = Yes, 0 = No, NA = assessment could not be made.
centroid_type: The type of centroid. Main refers to centroid of the largest polygon.  BB refers to the polygon bounding box.
relative_distance: Relative distance between the occurrence point and the nearest centroid (see note)
centroid_distance: Distance (decimal degrees) between the occurrence point and the nearest centroid.
dist_max: Furthest distance from the centroid within the political division.
}
\description{
This function is designed to use the BIEN centroid detection method which provides both relative and absolute distances to know centroids.  The lower the relative distance, the more likely that the occurrence is a centroid.
}
\note{
Relative centroid distance is calculated as: (distance between the occurrence record and centroid)/(distance between centroid and the furthest from it within the region)
}
\examples{

# Load required libraries

library(occTest)
library(piggyback)
library(sf)
library(tidyverse)

# Create temporary directory to store the centroid data

temp_dir <- tempdir()

# Download the data from Github

pb_download(file = "centroids_long_format.gpkg",
           repo = "pepbioalerts/occTest",
           tag = "data",
           dest = temp_dir,
           overwrite = TRUE)

# Load centroid file

centroids <- st_read(file.path(temp_dir,"centroids_long_format.gpkg"))

# Load example data set

occData <- read.csv(system.file('ext/exampleOccData.csv',
                               package = 'occTest')) |>
 filter(!is.na(MAPX) & !is.na(MAPY) ) |>
 st_as_sf(coords = c("MAPX", "MAPY")) |>
 mutate(country  = COUNTRYRECORD)


# Run function and append to existing data

occData |>
 bind_rows(  centroid_assessment_BIEN(occurrences = occData,
                                      centroid_data = centroids))
}
\author{
Brian Maitner & Pep Serra-Diaz
}
