% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{centroidDetection}
\alias{centroidDetection}
\title{Centroid detection function}
\usage{
centroidDetection(df, xf, yf, cf, method = "all", do = TRUE, verbose = FALSE)
}
\arguments{
\item{df}{\emph{data}.frame of species occurrences}

\item{xf}{\emph{character}. column name in df containing the x coordinates}

\item{yf}{\emph{character}. column name in df containing the y coordinates}

\item{cf}{\emph{character}. Country field in the species data.frame (df)}

\item{method}{\emph{character} Vector with the methods to detect centroids}

\item{do}{\emph{logical} Should range analysis be performed? Default TRUE}

\item{verbose}{\emph{character} Print messages? Default FALSE}
}
\value{
\emph{data.frame}
}
\description{
Identify occurrence records located near centroids.
}
\details{
This is a wrapper function that encompasses different methods

Current methods implemented for centroid detection are 'BIEN' (uses iterative procedure with threshold selection of distance to centroid)\cr
, and method 'CoordinateCleaner' implementing methods used in \link[CoordinateCleaner]{CoordinateCleaner-package} package.
}
\examples{
sp_file <- system.file('ext/exampleOccData.csv',package='occTest')
occ_species_example <- read.table(file = sp_file,header = T,sep = ',',as.is = T)
occ <- occ_species_example[1:50,c('MAPX','MAPY','COUNTRYRECORD')]
centroidDetection(df=occ,xf='MAPX',yf = 'MAPY',cf = 'COUNTRYRECORD')
}
\seealso{
\link[CoordinateCleaner]{cc_cap} \link[CoordinateCleaner]{cc_cen}

Other analysis: 
\code{\link{.nearestcell3}()},
\code{\link{countryStatusRangeAnalysis}()},
\code{\link{duplicatesexcludeAnalysis}()},
\code{\link{geoEnvAccuracy}()},
\code{\link{humanDetection}()},
\code{\link{timeAccuracy}()}
}
\author{
Brian Maitner, Josep M Serra-Diaz, Cory Merow. Functions from CoordinateCleaner implemented by A Zizka (CoordinateCleaner)
}
\concept{analysis}
\keyword{internal}
