% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{centroidDetection}
\alias{centroidDetection}
\title{Centroid detection function}
\usage{
centroidDetection(df = dat, xf = x.field, yf = y.field,
  cf = c.field, idf = taxonobservation.id, .ntv.ctry = ntv.ctry,
  .inv.ctry = inv.ctry, .points.proj4string = points.proj4string,
  .r.env = r.env, .countries.shapefile = countries.shapefile,
  cfsf = countryfield.shapefile, method = "all", do = T,
  verbose = F)
}
\arguments{
\item{df}{Data.frame of species occurrences}

\item{xf}{the field in the dataframe containing the x cordinates}

\item{yf}{the field in the dataframe containing the y cordinates}

\item{.ntv.ctry}{Native country in ISO3}

\item{.inv.ctry}{Invasive country in ISO3}

\item{.points.proj4string}{Proj4string argument}

\item{.r.env}{R env}
}
\value{
list
}
\description{
Centroid detection function
}
\examples{
\dontrun{
example<-"goes here"
}
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{internal}
