% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{countryStatusRangeAnalysis}
\alias{countryStatusRangeAnalysis}
\title{Range analysis}
\usage{
countryStatusRangeAnalysis(df = dat, xf = x.field, yf = y.field,
  .ntv.ctry = ntv.ctry, .inv.ctry = inv.ctry, .c.field = c.field,
  .points.proj4string = points.proj4string,
  .countries.shapefile = countries.shapefile,
  cfsf = countryfield.shapefile, excludeUnknownRanges = F,
  excludeNotmatchCountry = F, doRangeAnalysis = T, verbose = F)
}
\arguments{
\item{df}{Data.frame of species occurrences}

\item{xf}{the field in the dataframe containing the x cordinates}

\item{yf}{the field in the dataframe containing the y cordinates}

\item{.ntv.ctry}{Native country in ISO3}

\item{.inv.ctry}{Invasive country in ISO3}

\item{.c.field}{Country field in the species dataframe}

\item{.points.proj4string}{Proj4string for dataframe}

\item{.countries.shapefile}{spatialPolygonDataFrame indicating countries}

\item{cfsf}{field of the country spatialPolygonDataFrame indicaing the country in ISO3}
}
\value{
list
}
\description{
Range analysis of species records based on invsive and native country
}
\examples{
\dontrun{
example<-"goes here"
}
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{internal}
