% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_nativeStatusCtry.R
\name{ctryToIso3}
\alias{ctryToIso3}
\title{Convert country names to ISO3 codes}
\usage{
ctryToIso3(x, method = "countrycode")
}
\arguments{
\item{x}{character. country name}
}
\description{
using package countrycode
}
\details{
right now not implemented with fuzzy matching, but is case insensitive.
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{internal}
