% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nativeStatusCtry.R
\name{ctryToIso3}
\alias{ctryToIso3}
\title{Convert country names to ISO3 codes}
\usage{
ctryToIso3(x, method = "countrycode")
}
\arguments{
\item{x}{character. country name}

\item{method}{character. Package name used to derive IS03 codes. Options are 'countrycode' (default) or 'GNRS'.}
}
\description{
Froma character it uses different methods to derive country ISO3 digit codes
}
\details{
right now not implemented with fuzzy matching, but is case insensitive. Methods implemented, 'countrycode' and 'GNRS'
}
\seealso{
\link[countrycode]{countrycode} and \link[GNRS]{GNRS}

Other spStatus: 
\code{\link{nativeStatusCtry}()}
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr) (adpatation) \cr 
\link[countrycode]{countrycode-package} Vincent Arel-Bundock vincent.arel-bundock@umontreal.ca ,\cr
\link[GNRS]{GNRS}  Brad Boyle, Brian Maitner
}
\concept{spStatus}
\keyword{spStatus}
