% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_defaultSettings.R
\name{defaultSettings}
\alias{defaultSettings}
\title{load default settings for occTest}
\usage{
defaultSettings()
}
\value{
list of lists with all different parameters to use in occProfile function
}
\description{
Loads a list of lists with the different default parameters for analysis, outputs and grading needed in occTest
}
\details{
it can be use internally or it can be used by a user to subsequently modify parameters
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{user}
