% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geo.R
\name{.coords2country}
\alias{.coords2country}
\title{Extracts country ISO3 based on locations}
\usage{
.coords2country(
  xydat,
  xf = xf,
  yf = yf,
  .countries.shapefile = NULL,
  .points.crs = NULL,
  ctryNameField = NULL,
  verbose = FALSE
)
}
\arguments{
\item{xydat}{\emph{data.frame} with x and y coordinates}

\item{xf}{\emph{character} name of the x coordinate field}

\item{yf}{\emph{character} name of the y coordinate field}

\item{.countries.shapefile}{\emph{multipolygon} sf objects  of world countries and their associated ISO3 codes}

\item{.points.crs}{\emph{character} crs for the occurrence data}

\item{ctryNameField}{\emph{character} Column name in .countries shapefile where ISO3 are indicated}

\item{verbose}{\emph{logical} Print messages?}
}
\value{
\emph{character} factor vector with ISO3 codes for countries
}
\description{
Extracts country ISO3 based on locations
}
\author{
Josep M Serra Diaz
}
\concept{Geo}
