% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Geo.R
\name{.getSRTM}
\alias{.getSRTM}
\title{Extracts a dataframe based on dataframe of xy locations}
\usage{
.getSRTM(xydat, download = T, path = getwd(), verbose = F)
}
\arguments{
\item{xydat}{A dataframe with x and y coordinates}

\item{download}{Default to T. Whether the data should be downloaded}

\item{path}{where the downloads should go. Default to the currendt directory}

\item{verbose}{if you want to print messages of progress or warnings}
}
\value{
Factor with ISO3 codes for countries
}
\description{
Based on getData from raster
}
\examples{
\dontrun{

}

}
\seealso{
Other Geo: \code{\link{.coords2country}}
}
\concept{Geo}
