% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{.getSRTM}
\alias{.getSRTM}
\title{Download SRTM elevation raster}
\usage{
.getSRTM(xydat, download = TRUE, path = tempdir(), verbose = FALSE)
}
\arguments{
\item{xydat}{A dataframe with x and y coordinates}

\item{download}{Default to TRUE. Whether the data should be downloaded}

\item{path}{where the downloads should go. Default to the current directory}

\item{verbose}{if you want to print messages of progress or warnings}
}
\value{
raster
}
\description{
Download SRTM elevation raster
}
\details{
Basedd on getData from raster
}
\note{
borrowed from raster package but adapted to work directly within the occTest workflow
}
\seealso{
\link[raster]{getData}

Other Geo: 
\code{\link{.coords2country}()}
}
\concept{Geo}
