% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_OtherFuncs.R
\name{.hijack}
\alias{.hijack}
\title{Hijack functions}
\usage{
.hijack(FUN, ...)
}
\value{
character
}
\description{
Hijacking functions to rename them
}
\details{
In the occTest package this is used to get to a same function name for different OS implementing differnt parallelization systems
}
\note{


#Internal function that I got from stack overflow MrFlick
}
\examples{
\dontrun{
.data.frame <- hijack(data.frame, stringsAsFactors = FALSE)
dat <- .data.frame(x1 = 1:3, x2 = c("a", "b", "c"))
str(dat)  # yay! strings are character
}
}
\references{

}
\seealso{


Other internal: 
\code{\link{.find_and_transform_dates}()},
\code{\link{.get_os}()},
\code{\link{.gimme.score}()},
\code{\link{lazylogic}()},
\code{\link{rm.all}()},
\code{\link{splitSpname}()}
}
\author{

}
\concept{internal}
\keyword{internal}
