% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_OtherFuncs.R
\name{.multiple.strsplit}
\alias{.multiple.strsplit}
\title{multiple splitting of strings}
\usage{
.multiple.strsplit(x, multiple.splits)
}
\arguments{
\item{x}{character.}

\item{multiple.splits}{character. Vector with the characters used to recursively split x}
}
\value{
character
}
\description{
uses several splitting characters to split a character string.
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
Pep Serra-Diaz
}
\concept{}
\keyword{internal}
