% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{.nearestcell3}
\alias{.nearestcell3}
\title{SeaLand reassignment}
\usage{
.nearestcell3(dat, rst, xf, yf, verbose = FALSE)
}
\arguments{
\item{dat}{\emph{data.frame} of species occurrences}

\item{rst}{\emph{SpatRast} class object with 1 being the target habitat (land or sea)}

\item{xf}{the column name in dat containing the x coordinators}

\item{yf}{the column name in dat containing the y coordinators}

\item{verbose}{\emph{logical}. Print messages? Default FALSE}
}
\value{
\emph{list}
}
\description{
Reassign coastal coordinates as needed to be in
}
\details{
function inspired in nearestcell in biogeo but modified
}
\seealso{
original code at \link[biogeo]{nearestcell}

Other analysis: 
\code{\link{centroidDetection}()},
\code{\link{countryStatusRangeAnalysis}()},
\code{\link{duplicatesexcludeAnalysis}()},
\code{\link{geoEnvAccuracy}()},
\code{\link{humanDetection}()},
\code{\link{timeAccuracy}()}
}
\author{
Mark Robertson and Veron Visser (original \link[biogeo]{biogeo-package}), Josep M Serra-Diaz (modifs)
}
\concept{analysis}
\keyword{internal}
