% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{.nearestcell3}
\alias{.nearestcell3}
\title{SeaLand reassignment}
\usage{
.nearestcell3(dat, rst, xf, yf, verbose = FALSE)
}
\arguments{
\item{dat}{Data.frame of species occurrences}

\item{rst}{raster class object.}

\item{xf}{the field in the data frame containing the x coordinators}

\item{yf}{the field in the data frame containing the y coordinators}

\item{verbose}{logical. Print messages? Default FALSE}
}
\value{
list
}
\description{
SeaLand reassignment
}
\details{
(function inspired in nearestcell in biogeo but modified
}
\seealso{
original code at \link[biogeo]{nearestcell}

Other analysis: 
\code{\link{centroidDetection}()},
\code{\link{countryStatusRangeAnalysis}()},
\code{\link{duplicatesexcludeAnalysis}()},
\code{\link{geoEnvAccuracy}()},
\code{\link{humanDetection}()}
}
\author{
Mark Robertson and Veron Visser (original \link[biogeo]{biogeo-package}), Josep M Serra-Diaz (modifs)
}
\concept{analysis}
\keyword{internal}
