% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occOutliers_modif.R, R/outliers.R
\name{.presPercentile}
\alias{.presPercentile}
\title{omit outlying pres}
\usage{
.presPercentile(
  xy,
  percent = 95,
  unin = c("m", "km"),
  unout = c("ha", "km2", "m2")
)

.presPercentile(
  xy,
  percent = 95,
  unin = c("m", "km"),
  unout = c("ha", "km2", "m2")
)
}
\arguments{
\item{xy}{x y coordinates. Class SpatialPoints}

\item{percent}{numeric. Defaults to 95. Percentage to def object}

\item{unin}{character vector.}

\item{unout}{character vector.}
}
\description{
Divide raster by the sum of all cells.
}
\author{
Josep M Serra Diaz
}
\keyword{internal}
