% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.status.tracker.and.escaping}
\alias{.status.tracker.and.escaping}
\title{Workflow status tracker}
\usage{
.status.tracker.and.escaping(
  dataset.to.continue,
  wfo,
  wso,
  xf,
  yf,
  od,
  rsd,
  obf,
  sp,
  verbose = FALSE,
  as,
  ws,
  ts
)
}
\arguments{
\item{dataset.to.continue}{A dataframe containing occurrence data for checking.}

\item{wfo}{write full output}

\item{wso}{write simple output}

\item{xf}{The dataframe field containing the x values (e.g. "longitude")}

\item{yf}{The dataframe field containing the y values (e.g. "latitude")}

\item{od}{The output directory to use}

\item{rsd}{logical. Return spatial data?}

\item{obf}{Output base filename}

\item{sp}{character. Name of the species}

\item{verbose}{logical. Print messages? Defaults to FALSE}

\item{as}{list. analysis settings}

\item{ws}{list. writoutsettings}

\item{ts}{tlist. able setting}
}
\value{
Original dataframe, dat.  Used primarily to generate warning messages.
}
\description{
Track status and write useful output
}
\seealso{
Other checks: 
\code{\link{.addmainfields2}()},
\code{\link{.check.geospatial.data}()},
\code{\link{.checkdatastr2}()},
\code{\link{.checkfields}()}
}
\author{
Josep M Serra Diaz
}
\concept{checks}
