% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_OtherFuncs.R
\name{.subsetlist.isNA}
\alias{.subsetlist.isNA}
\title{Subset to non-NA lists}
\usage{
.subsetlist.isNA(x)
}
\arguments{
\item{x}{List}
}
\value{
list
}
\description{
Subset a list removing NA entries
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{

}
\concept{}
\keyword{internal}
