% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{duplicatesexcludeAnalysis}
\alias{duplicatesexcludeAnalysis}
\title{Duplicated records}
\usage{
duplicatesexcludeAnalysis(
  df = dat,
  xf = x.field,
  yf = y.field,
  resolution.in.minutes = res.in.minutes,
  raster.grid = NULL,
  verbose = F
)
}
\arguments{
\item{df}{Data.frame of species occurrences}

\item{xf}{the field in the dataframe containing the x cordinates}

\item{yf}{the field in the dataframe containing the y cordinates}

\item{resolution.in.minutes}{the resolution of environmental data used, specified in minutes}

\item{raster.grid}{An optional raster grid}
}
\value{
list of three components: Dups.Exact = exact duplicate records, Dups.Grid= duplicates within environmental gridcell, continue = dataframe with good records
}
\description{
Duplicated records
}
\details{
(inspired by biogeo duplicatesexclude function)
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{internal}
