% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{envOutliers}
\alias{envOutliers}
\title{ENVIRORNMENTAL OUTLIERS (using both jacknife and boxplot)}
\usage{
envOutliers(.r.env = r.env, df = dat, xf = x.field, yf = y.field,
  .th.perc.outenv = th.perc.outenv, .sp.name = sp.name,
  method = "all", .projString = points.proj4string, do = T,
  verbose = F)
}
\arguments{
\item{.r.env}{R enviroment}

\item{df}{Data.frame of species occurrences}

\item{xf}{the field in the dataframe containing the x cordinates}

\item{yf}{the field in the dataframe containing the y cordinates}

\item{.th.perc.outenv}{Something}

\item{.sp.name}{Species name}

\item{outlier.method}{Method to use for detecting outliers.  Defaults to 'all'}
}
\value{
list
}
\description{
Detect environmental outliers using jacknife and boxplot
}
\examples{
\dontrun{
example<-"goes here"
}
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{internal}
