% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{filterMissing}
\alias{filterMissing}
\title{Check for missing coordinates}
\usage{
filterMissing(df, xf = x.field, yf = y.field, verbose = F)
}
\arguments{
\item{df}{Data.frame of species occurrences}

\item{xf}{the field in the dataframe containing the x cordinates}

\item{yf}{the field in the dataframe containing the y cordinates}
}
\value{
list with two dataframes: stay = coordinates missing and continue = occurrence that you can retain for further analysis
}
\description{
checks for missing coordinates in the occurrence dataframe
}
\details{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{internal}
