% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{filterMissing}
\alias{filterMissing}
\title{Check for missing coordinates}
\usage{
filterMissing(df, xf, yf, verbose = FALSE)
}
\arguments{
\item{df}{\emph{data.frame} of species occurrences}

\item{xf}{\emph{character}. The field in the data.frame containing the x coordinates}

\item{yf}{\emph{character}. The field in the data.frame containing the y coordinates}

\item{verbose}{\emph{logical}. Print messages? Default FALSE}
}
\value{
\emph{list} with two \emph{data.frame}: stay = coordinates missing and continue = occurrence that you can retain for further analysis
}
\description{
checks for missing coordinates in the occurrence dataframe
}
\examples{
ds <- data.frame (x=c(runif (n = 100),NA),y=c(runif (n = 100),1000))
filterMissing(ds,xf='x',yf='y')
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{internal}
