% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{filterMissing}
\alias{filterMissing}
\title{Check for missing coordinates}
\usage{
filterMissing(df, xf, yf, verbose = FALSE)
}
\arguments{
\item{df}{dat.frame of species occurrences}

\item{xf}{character. The field in the data.frame containing the x coordinates}

\item{yf}{character. The field in the data.frame containing the y coordinates}

\item{verbose}{logical. Print messages? Default FALSE}
}
\value{
List with two dataframes: stay = coordinates missing and continue = occurrence that you can retain for further analysis
}
\description{
checks for missing coordinates in the occurrence dataframe
}
\examples{
k <- data.frame (x=c(runif (n = 100),NA),y=c(runif (n = 100),1000))
filterMissing(k,xf='x',yf='y')
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{internal}
