% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Unsorted_Functions.R
\name{findEnvOutliers}
\alias{findEnvOutliers}
\title{Find outlying occurrence data in environmental space}
\usage{
findEnvOutliers(myPres, myEnv = NULL, pvalSet = 1e-05,
  checkPairs = F, verbose = T)
}
\arguments{
\item{myPres}{a `SpatialPointsDataFrame`}

\item{pvalSet}{numeric; p-value used in Grubb's test for outlier (see package `outliers`)}

\item{env}{a `RasterStack` of env layers. If NULL, it is assumed that `myPres` is a data.frame of the environmental values (columns) at presence locations (rows)}
}
\value{
Returns a list of SpatialPointsDataFrames with (1) good presence points (2) spatial outliers and (3) environmental outliers.
}
\description{
Environmental outliers
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
