% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occOutliers_modif.R
\name{findEnvOutliers}
\alias{findEnvOutliers}
\title{Find outlying occurrence data in environmental space}
\usage{
findEnvOutliers(
  pres,
  pvalSet = 1e-05,
  method = "grubbs",
  checkPairs = TRUE,
  kRosner = NULL
)
}
\arguments{
\item{pres}{an `sf` points object describing the locations of species records.}

\item{pvalSet}{user-specified p-value for assessing the significance of Grubbs test statistic.}

\item{method}{character; options are 'iqr', 'grubbs', 'dixon', 'rosner'}

\item{checkPairs}{logical; check for a single pair of outliers using the Grubbs test. This can only be performed for sample sizes <30. Only a single test is used because repeating it tends to throw out more points than seem reasonable, by eye. The value has no effect unless `method='grubbs'`.}

\item{kRosner}{integer between 1 and 10. Determines the number of outliers suspected with a Rosner test. The value has no effect unless `method='rosner'`.}
}
\value{
Returns the indices of environmental outliers
}
\description{
Environmental outliers
}
\details{
See Examples.
}
\examples{
#select coordinates only
myPres=read.csv(system.file('ext/exampleOccData.csv',
                            package='occTest'))
#select coordinates only
myPres <- myPres[,2:3]
myPres <- myPres[complete.cases(myPres),]
#extract values
myEnv <- terra::rast(system.file('ext/AllEnv.tif',package='occTest'))
myPres <- sf::st_as_sf(myPres,coords = c(1,2))
myPresDF <- cbind (myPres,data.frame(terra::extract(myEnv,myPres)))
#find outliers
presOut=findEnvOutliers(pres=myPresDF,pvalSet=1e-5)
}
\author{
Cory Merow <cory.merow@gmail.com>
}
