% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{findSpatialOutliers}
\alias{findSpatialOutliers}
\title{Find outlying occurrence data in geographical space}
\usage{
findSpatialOutliers(myPres, pvalSet = 1e-05, checkPairs = T, verbose = T)
}
\arguments{
\item{myPres}{raster* object}

\item{pvalSet}{numeric. p value set to identify outliers}

\item{checkPairs}{logical.}

\item{verbose}{logical. print messages}
}
\description{
Divide raster by the sum of all cells.
}
\author{
C Merow
}
\keyword{internal}
