% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Unsorted_Functions.R
\name{findSpatialOutliers}
\alias{findSpatialOutliers}
\title{omit outlying pres}
\usage{
findSpatialOutliers(myPres, pvalSet = 1e-05, checkPairs = T, verbose = T)
}
\arguments{
\item{x}{raster* object}

\item{...}{additional functions to be passed to \code{\link[raster]{writeRaster}}}
}
\description{
Divide raster by the sum of all cells.
}
