% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occOutliers_modif.R
\name{findSpatialOutliers}
\alias{findSpatialOutliers}
\title{Find outlying occurrence data in geographic space}
\usage{
findSpatialOutliers(
  pres,
  pvalSet = 1e-05,
  method = "grubbs",
  checkPairs = FALSE,
  kRosner = NULL
)
}
\arguments{
\item{pres}{a `sf` points object describing the locations of species records.}

\item{pvalSet}{user-specified p-value for assessing the significance of Grubbs test statistic.}

\item{method}{character; options are 'iqr', 'grubbs', 'dixon', 'rosner'}

\item{checkPairs}{logical; check for a single pair of outliers using the Grubbs test. This can only be performed for sample sizes <30. Only a single test is used because repeating it tends to throw out more points than seem reasonable, by eye. The value has no effect unless `method='grubbs'`.}

\item{kRosner}{integer between 1 and 10. Determines the number of outliers suspected with a Rosner test. The value has no effect unless `method='rosner'`.}
}
\value{
Returns the indices of spatial outliers
}
\description{
Spatial outliers
}
\details{
See Examples.
}
\examples{
myPres=read.csv(system.file('./ext/exampleOccData.csv',
                            package='occTest')) |> dplyr::select(c('MAPX','MAPY'))
myPres = sf::st_as_sf(myPres[complete.cases(myPres),],
                      coords =c('MAPX','MAPY'))
presOut=occTest::findSpatialOutliers(pres=myPres, pvalSet=1e-5)

}
\author{
Cory Merow <cory.merow@gmail.com>
}
