% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{geoEnvAccuracy}
\alias{geoEnvAccuracy}
\title{Coordinate accuracy}
\usage{
geoEnvAccuracy(df, xf = x.field, yf = y.field, af = a.field,
  dsf = ds.field, ef = e.field, method = "all", r.env,
  accept.threshold.cell = 0.5, accept.threshold.env = 0.5,
  bearing.classes = 10, distance.classes = 5, env.quantiles = c(0.3,
  0.7), elev.threshold = 100, raster.elevation = NULL, verbose = F,
  do = do.geoEnvAccuracy)
}
\arguments{
\item{df}{Data.frame of species occurrences}

\item{xf}{the field in the dataframe containing the x cordinates}

\item{yf}{the field in the dataframe containing the y cordinates}

\item{af}{the field where the geographic uncertainty is (in the same )}

\item{.r.env}{R enviroment}

\item{accept.threshold}{acceptance threshold for how much percentage of the Area of uncertainty in the cell we want to accept}
}
\description{
Detect environmental outliers using jacknife and boxplot
}
\examples{
\dontrun{
example<-"goes here"
}
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{Analysis}
