% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{geoOutliers}
\alias{geoOutliers}
\title{Detect geographic outliers}
\usage{
geoOutliers(
  df,
  xf,
  yf,
  .alpha.parameter,
  .distance.parameter = 1000,
  .medianDeviation.parameter = 5,
  .samplingIntensThreshold.parameter = 0.1,
  .mcp_percSample = 95,
  method = "all",
  .points.crs,
  do = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{\emph{data.frame}. Species occurrences data with x and y coordinates}

\item{xf}{\emph{character}. Column name in df containing the x coordinates}

\item{yf}{\emph{character}. Column name in df containing the y coordinates}

\item{.alpha.parameter}{\emph{numeric}. Parameter setting for alphahull. Alpha parameter recommended is 2. No default}

\item{.distance.parameter}{\emph{numeric}. Maximum distance allowed. Default to 1000}

\item{.medianDeviation.parameter}{\emph{numeric}. Deviation parameter to . Default to 0.1}

\item{method}{\emph{character}. Vector of methods to be used. See details. Default 'all'}

\item{.points.crs}{\emph{character}. crs coding indicating coordinate reference system}

\item{do}{\emph{logical}.. Should tests be performed? Default TRUE}

\item{verbose}{\emph{logical}.. Print messages? Default FALSE}
}
\value{
a \emph{data.frame} with the results of the tests
}
\description{
Detect geographical outliers using several tests
}
\details{
Methods implented are 'alphaHull', detecting species outside an alphaHull level (default 2), \cr
'alphaHull' \cr
'distance' corresponds to 'distance' method in  CoordinateCleaner::cc_outl(method='distance'). See ?CoordinateCleaner::cc_outl \cr
'median' corresponds to 'mad' method in  CoordinateCleaner::cc_outl(method='mad'). See ?CoordinateCleaner::cc_outl \cr
'quantSamplingCorrected' corresponds to 'mad' method in  CoordinateCleaner::cc_outl(method='quantile'). See ?CoordinateCleaner::cc_outl \cr
'grubbs' implements Grubbs test to find spatial outliers. See ?findSpatialOutliers for details \cr
}
\examples{
\dontrun{
sp_file <- system.file('ext/exampleOccData.csv',package='occTest')
occ <- read.table(file = sp_file,header = TRUE,sep = ',',as.is = TRUE,nrows = 50)
occ <- occ[,c('MAPX','MAPY')]
occ <- occ[stats::complete.cases(occ),]
geoOutliers(df=occ,xf = 'MAPX',yf = 'MAPY',.alpha.parameter = 2)
}
}
\seealso{
getPointsOutAlphaHull, \link[CoordinateCleaner]{cc_outl}, findSpatialOutliers
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr), Cory Merow (cmerow@gmail.com)
}
\keyword{internal}
