% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Analysis.R
\name{geoOutliers}
\alias{geoOutliers}
\title{Detect geographic outliers}
\usage{
geoOutliers(
  df = dat,
  xf = x.field,
  yf = y.field,
  .alpha.parameter = alpha.parameter,
  .distance.parameter = 1000,
  .medianDeviation.parameter = 5,
  .samplingIntensThreshold.parameter = 0.1,
  .projString = points.proj4string,
  method = "all",
  do = T,
  verbose = F
)
}
\arguments{
\item{df}{Data.frame of species occurrences}

\item{xf}{the field in the dataframe containing the x coordinates}

\item{yf}{the field in the dataframe containing the y coordinates}

\item{.alpha.parameter}{parameter setting for alphahull}

\item{.distance.parameter}{numeric. Maximum distance allowed. Default to 1000}

\item{.medianDeviation.parameter}{numeric. Deviation parameter to . Default to 0.1}

\item{.projString}{proj4string character. Indicate coordinate reference system}

\item{method}{charcter. Vector of methods to be used. See details. Default 'all'}

\item{do}{logical. Should tests be performed? Default T}

\item{verbose}{logical. Print messages? Default F}
}
\value{
data.frame
}
\description{
Detect geographic outliers
}
\details{
Methods implented are 'alphaHull', detecting species outside an alphaHull level (default 2), \cr
'alphaHull' \cr
'distance' corresponds to 'distance' method in  CoordinateCleaner::cc_outl(method='distance'). See ?CoordinateCleaner::cc_outl \cr
'median' corresponds to 'mad' method in  CoordinateCleaner::cc_outl(method='mad'). See ?CoordinateCleaner::cc_outl \cr
'quantSamplingCorrected' corresponds to 'mad' method in  CoordinateCleaner::cc_outl(method='quantile'). See ?CoordinateCleaner::cc_outl \cr
'grubbs' implements Grubbs test to find spatial outliers. See ?findSpatialOutliers for details \cr
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{
getPointsOutAlphaHull, CoordinateCleaner::cc_outl, findSpatialOutliers
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{internal}
