% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{humanDetection}
\alias{humanDetection}
\title{Hyper Human Influence}
\usage{
humanDetection(
  df,
  xf,
  yf,
  method = "all",
  ras.hii,
  .th.human.influence,
  .points.crs = NULL,
  do = TRUE,
  verbose = FALSE,
  output.dir = NULL
)
}
\arguments{
\item{df}{\emph{data.frame} of species occurrences}

\item{xf}{\emph{character} the field in the data frame containing the x coordinates}

\item{yf}{\emph{character} the field in the data frame containing the y coordinates}

\item{method}{\emph{character}. Indicate which tests to use. Default 'all'. See Details}

\item{ras.hii}{\emph{spatRaster} of human influence index}

\item{.th.human.influence}{\emph{numeric} threshold of human influence index}

\item{.points.crs}{\emph{crs} argument for df}

\item{do}{\emph{logical}. Should range analysis be performed? Default TRUE}

\item{verbose}{\emph{logical} Print messages? Default FALSE}

\item{output.dir}{\emph{character}. Output directory}
}
\value{
a \emph{data.frame} with the results of the tests
}
\description{
Detect occurrences in heavily human-impacted environments
}
\details{
It uses several methods to detect records in high human influence records.\cr
Current implemented methods are: \cr
'hii' using a raster and a threshold of human influence
'urban' using a layer of urban areas. Method implemented in CoordinateCleaner package.
}
\examples{
\dontrun{
sp_file <- system.file('ext/exampleOccData.csv',package='occTest')
occ_species_example <- read.table(file = sp_file,header = TRUE,sep = ',',as.is = TRUE)
occ <- occ_species_example[1:50,c('MAPX','MAPY','COUNTRYRECORD')]
raster_humanImpact <- terra::rast(xmin=min(occ$MAPX)-1,
                                  xmax=max(occ$MAPX)+1,
                                  ymin=min(occ$MAPY)-1,
                                  ymax=max(occ$MAPY)+1,
                                  crs="EPSG:4326")
set.seed(33)
terra::values(raster_humanImpact) <- round(runif(terra::ncell(raster_humanImpact),
                                                 min = 0,max=100),digits = 0)
humanDetection(df=occ, xf='MAPX',yf = 'MAPY',ras.hii = raster_humanImpact,.th.human.influence = 50)
}
}
\seealso{
\link[CoordinateCleaner]{cc_urb} for CoordinateCleaner functions

Other analysis: 
\code{\link{.nearestcell3}()},
\code{\link{biogeo_outliers}()},
\code{\link{centroidDetection}()},
\code{\link{countryStatusRangeAnalysis}()},
\code{\link{duplicatesexcludeAnalysis}()},
\code{\link{geoEnvAccuracy}()},
\code{\link{timeAccuracy}()}
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr). A Zizka (CoordinateCleaner functions)
}
\concept{analysis}
\keyword{internal}
