% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{landUseSelect}
\alias{landUseSelect}
\title{Selection of records within a specified land use}
\usage{
landUseSelect(
  df,
  xf,
  yf,
  method = "in",
  .points.crs = NULL,
  ras.landUse,
  .landUseCodes,
  do = TRUE,
  verbose = FALSE,
  output.dir
)
}
\arguments{
\item{df}{\emph{data.frame}. Dataset of species occurrences with their coordinates and other information.}

\item{xf}{\emph{character}. column name in df containing the x coordinates}

\item{yf}{\emph{character}. column name in df containing the y coordinates}

\item{method}{\emph{character}. Select 'in' (default) when records are selected when inside specified land use types (type 'out' otherwise)}

\item{.points.crs}{\emph{character}. crs indicating coordinate reference system. Default to null, which assumes ESPG:4326}

\item{ras.landUse}{\emph{SpatRaster} Land use raster with integer codes for land use classes.}

\item{.landUseCodes}{\emph{numeric}. Vector of specified land use codes to either select records (method 'in') or to exclude records (method 'out')}

\item{do}{\emph{logical}. Should range analysis be performed?}

\item{verbose}{\emph{logical}. Print messages? Default  FALSE}

\item{output.dir}{\emph{logical}. Output directory}
}
\value{
a \emph{data.frame} with the results of the tests
}
\description{
Selection of records within a specified land use
}
\examples{
\dontrun{
sp_file <- system.file('ext/exampleOccData.csv',package='occTest')
occ <- read.table(file = sp_file,header = TRUE,sep = ',',as.is = TRUE,nrows = 50)
raster_landUse <- terra::rast(xmin=min(occ$MAPX)-1,
                              xmax=max(occ$MAPX)+1,
                              ymin=min(occ$MAPY)-1,
                              #' ymax=max(occ$MAPY)+1,
                              crs="EPSG:4326")
terra::values(raster_landUse) <- c(rep (1,terra::ncell(raster_landUse)/2),
                                   rep (2,terra::ncell(raster_landUse)/2))
landUseSelect(df=occ,xf = 'MAPX',yf = 'MAPY',ras.landUse = raster_landUse,.landUseCodes = 1)                                                  
}

}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{internal}
