% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFuncs.R
\name{lazylogic}
\alias{lazylogic}
\title{a lazy logic omitting NAs}
\usage{
lazylogic(e)
}
\arguments{
\item{e}{expression}
}
\value{
logic
}
\description{
evaluates a conditional expression where if any element is NA the output is FALSE instead of NA
}
\seealso{
Other internal: 
\code{\link{.get_os}()},
\code{\link{rm.all}()},
\code{\link{splitSpname}()}
}
\concept{internal}
