% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{mcp_test}
\alias{mcp_test}
\title{Find outlying occurrence data in geographical space from convex hulls}
\usage{
mcp_test(data, xcoord, ycoord, percentage = 95)
}
\arguments{
\item{data}{a data.frame  with coordinate information}

\item{xcoord}{character. Column with X coordinate information in data}

\item{ycoord}{character. Column with Y coordinate information in data}

\item{percentage}{numeric. Percentage of data used to build the minimum convex polygon}
}
\value{
Returns a list of SpatialPointsDataFrames with (1) good presence points (2) spatial outliers and (3) environmental outliers.
}
\description{
Geographic outliers
}
\details{
only works for WGS84 lat long
}
\author{
Coline CF Boonmann <colineboonman@bio.au.dk >
}
\keyword{internal}
