% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_nativeStatusCtry.R
\name{nativeStatusCtry}
\alias{nativeStatusCtry}
\title{Check native and alien ranges in countries}
\usage{
nativeStatusCtry(spName, resolveNative = T, resolveAlien = T, verbose = T)
}
\arguments{
\item{spName}{character. Species name in the form of Genus species}

\item{resolveNative}{logical. Should the function attempt to find countries where species are considered native?}

\item{resolveAlien}{logical. Should the function attempt to find countries where species are considered introduced?}

\item{verbose}{logical. Want to print information during the process?}
}
\value{
list with two vectors: ntvCtry and invCtry, showing the countries in ISO3 standard codes
}
\description{
performs several queries to
}
\details{
right now based on the package originatr. Future implementation may query more databases.
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{internal}
