% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_filter.R
\name{occFilter}
\alias{occFilter}
\title{Filter occurrence records}
\usage{
occFilter(
  df,
  by = "testBlock",
  errorAcceptance = "relaxed",
  errorThreshold = NULL
)
}
\arguments{
\item{df}{data.frame outputed from occurrenceClassify}

\item{level}{numeric. Level of application of the filtering process. 1 is the coarsest 3 is the minimum}

\item{threshold}{character. Phylosophy for filtering based on threshold. Option are majority, relaxed, stringent}

\item{customThresholds}{data.frame. Specify for each dimension of flags the specific threshold to use. It overides the parameters in thresholds. We recomend building that table based on the functio buildCustomThresholds.}
}
\value{
list of 2 data.frames
}
\description{
Filter occurrence records
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{filter}
