% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{occSimpFilter}
\alias{occSimpFilter}
\title{Runs tests and validates data}
\usage{
occSimpFilter(
  spOcc,
  env,
  speciesName = "My species",
  x = "x",
  y = "y",
  date = NULL,
  isoCountry = NULL,
  classification = "majority",
  filterCols = T
)
}
\arguments{
\item{spOcc}{data.frame. Object with the coordinate data.}

\item{env}{raster or rasterStack. Environmental data (e.g. typically climatic).}

\item{speciesName}{character. Name of the species.}

\item{x}{name of the field with the coordinate x. Default 'x'}

\item{y}{name of the field with the coordinate y. Default 'y'}

\item{date}{name of the field with the values of the reported timestamp of the record. Default NULL}

\item{isoCountry}{name of the field with the values of the reported country of the record. Default NULL}

\item{classification}{character. Indicates the thresholds philosophy applied to classify errors in occurrence data. Possible values 'strict','relaxed','custom'}

\item{filterCols}{logical. Should only the initial input columns be retained in the output (the filtered dataframe)?}
}
\value{
a list of two. First element is a data.frame with profiled occurrence records with their associated profiled labels. Second element is a dataframe with all outputs of the analysis implemented.
}
\description{
Runs tests and validates data
}
\note{
There are several parameters in the function. The majority of them can be adjusted, but we also provide default values. We recommend those default values if the user is to use the geospatial data included in the package.
}
\examples{
\dontrun{
example<-"goes here"
}
}
\author{
Pep Serra-Diaz
}
