% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{occSimpFilter}
\alias{occSimpFilter}
\title{Runs tests and validates data}
\usage{
occSimpFilter(
  spOcc,
  env,
  speciesName = "My species",
  x = "x",
  y = "y",
  date = NULL,
  isoCountry = NULL,
  classification = "majority",
  filterCols = TRUE
)
}
\arguments{
\item{spOcc}{\emph{data.frame}. Object with the coordinate data.}

\item{env}{\emph{SpatRaster} object. Raster of environmental data (e.g. typically climatic).}

\item{speciesName}{\emph{character}. Name of the target species.}

\item{x}{\emph{character}. Name of the field in the data frame containing the x coordinates}

\item{y}{\emph{character}. Name of the field in the data frame containing the y coordinates}

\item{date}{\emph{character}. Name of the field with the values of the reported timestamp of the record. Default NULL}

\item{isoCountry}{\emph{character}. Name of the field with the values of the reported country of the record. Default NULL}

\item{classification}{\emph{character}.. Indicates the thresholds philosophy applied to classify errors in occurrence data. Possible values 'strict','relaxed','custom'}

\item{filterCols}{\emph{logical}.. Should only the initial input columns be retained in the output (the filtered dataframe)?}
}
\value{
a \emph{list} of two. First element is a data.frame with profiled occurrence records with their associated profiled labels. Second element is a dataframe with all outputs of the analysis implemented.
}
\description{
Runs tests and validates data
}
\note{
The majority of function parameters can be adjusted but we  provide default values. \cr 
We recommend those default values if the user is to use the geospatial data included in the package.\cr
but this automatic implementation (occTest + occFilter) missses some analysis to increase speed.
}
\examples{
\dontrun{
### THIS IS A CUT DOWN  EXAMPLE 
### visit vignetteXtra-occTest for more info
#load environmental SpatRaster
library (sf)
library (occTest)
#load occurrence data
occData <- read.csv (system.file('ext/exampleOccData.csv',package = 'occTest'))
#load environmental SpatRaster
renv <- terra::rast (system.file('ext/AllEnv.tif',package = 'occTest'))
#run occSimpFilter
out = occSimpFilter (spOcc=occData,
                     env=renv,
                     x='MAPX',y='MAPY')
}
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
