% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_plot.R
\name{plot_occFilter}
\alias{plot_occFilter}
\title{Display the filtering process}
\usage{
plot_occFilter(
  occTest_df,
  occfilter_list = NULL,
  tableSettings = NULL,
  show_plot = T
)
}
\arguments{
\item{occTest_df}{A data.frame returned by  {\link[=occTest]{occTest}}, i.e. the unfiltered data.frame}

\item{occfilter_list}{Optional, a list returned by  {\link[=occFilter]{occTest}}, the result of the filtering by \code{occTest_df}}

\item{tableSettings}{list. Elements corresponding to different settings of the input occurrence table, used to create the \code{occTest_df} object, if null, try to use the default settings}

\item{show_plot}{Logical, should the plots be plotted ?}
}
\value{
list of ggplots objects, of varying length, depending on wether the filtering was done by testBlock or testType
}
\description{
Display the filtering process
}
\details{
If \code{occfilter_list} is provided, display how the occurences passed the different tests, otherwise only plot the coordinates filtering step
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{
{\link[=occFilter]{occFilter}}  , {\link[=occTest]{occTest}}  , the {\link[=ggplot2]{ggplot2}} package
}
\author{
J Borderieux (jeremy.borderieux@agroparistech.fr)
}
\concept{}
\keyword{filter}
\keyword{plot}
