% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFuncs.R
\name{rm.all}
\alias{rm.all}
\title{Remove all objects}
\usage{
rm.all()
}
\value{
a clear environment
}
\description{
removes all objects in the environment
}
\seealso{
Other internal: 
\code{\link{.get_os}()},
\code{\link{lazylogic}()},
\code{\link{splitSpname}()}
}
\concept{internal}
