% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_defaultSettings.R
\name{setTableNames}
\alias{setTableNames}
\title{set table names internally}
\usage{
setTableNames(
  x.field = NULL,
  y.field = NULL,
  t.field = NULL,
  l.field = NULL,
  c.field = NULL,
  e.field = NULL,
  a.field = NULL,
  ds.field = NULL,
  taxonobservation.id = NULL
)
}
\arguments{
\item{x.field}{character. Name of the x coordinate field.}

\item{y.field}{character. Name of the y coordinate field.}

\item{t.field}{character. Name of the timestamp field.}

\item{l.field}{character. Name of the locality field.}

\item{c.field}{character. Name of the country code field.}

\item{e.field}{character. Name of the eleveation field.}

\item{a.field}{character. Name of the accuracy field.}

\item{ds.field}{character. Name of the dataset identifier field.}

\item{taxonobservation.id}{character. Name of the taxon observartion id field.}
}
\value{
list
}
\description{
set table names internally
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{user}
