% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_defaultSettings.R
\name{setTestBlocks}
\alias{setTestBlocks}
\title{set the important parameters}
\usage{
setTestBlocks(geo = T, lu = T, env = T, time = T)
}
\value{
list
}
\description{
set the important parameters
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{user}
