% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_defaultSettings.R
\name{setTestBlocks}
\alias{setTestBlocks}
\title{Set the tests to run}
\usage{
setTestBlocks(geo = T, lu = T, env = T, time = T)
}
\value{
list
}
\description{
function used to select which groups of tests you want in occTest workflow
}
\details{
See occTest::showTests for further information on tests used in the packages
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{user}
