% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{setTestBlocks}
\alias{setTestBlocks}
\title{Set the tests to run}
\usage{
setTestBlocks(geo = TRUE, lu = TRUE, env = TRUE, time = TRUE)
}
\arguments{
\item{geo}{logical. Should this family of tests be performed?}

\item{lu}{logical. Should this family of tests be performed?}

\item{env}{logical. Should this family of tests be performed?}

\item{time}{logical. Should this family of tests be performed?}
}
\value{
list
}
\description{
function used to select which groups of tests you want in occTest workflow
}
\details{
You can turn off an entire type of tests altogether by modifying this seetings. See occTest::showTests for further information on tests used in the packages
}
\examples{
defaultSettings_analysis  <- setTestBlocks()
#now we turn off the block of tests related to land use
mySettings_analysis  <- setTestBlocks(lu=FALSE)
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{user}
