% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{setTestTypes}
\alias{setTestTypes}
\title{Set the tests to run}
\usage{
setTestTypes(
  countryStatusRange = T,
  centroidDetection = T,
  humanDetection = T,
  landUseType = T,
  institutionLocality = T,
  geoOutliers = T,
  envOutliers = T,
  geoenvLowAccuracy = T
)
}
\arguments{
\item{countryStatusRange}{logical. Should this test type be performed?}

\item{centroidDetection}{logical. Should this test type be performed?}

\item{humanDetection}{logical. Should this test type be performed?}

\item{landUseType}{logical. Should this test type be performed?}

\item{institutionLocality}{logical. Should this test type be performed?}

\item{geoOutliers}{logical. Should this test type be performed?}

\item{envOutliers}{logical. Should this test type be performed?}

\item{geoenvLowAccuracy}{logical. Should this test type be performed?}
}
\value{
list with user analysis settings
}
\description{
function used to select which types of tests you want in occTest workflow (analysisSet)
}
\details{
See occTest::showTests for further information on tests used in the packages
}
\examples{
\dontrun{
defaultSettings_analysis <- setTestTypes()
#now we do not want to perform centroid geoenironmental accuracy type of tests
mySettings_analysis <- setTestTypes(geoenvLowAccuracy=F)
}
}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\keyword{user}
