% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_defaultSettings.R
\name{setTestTypes}
\alias{setTestTypes}
\title{Set the tests to run}
\usage{
setTestTypes(
  countryStatusRange = T,
  centroidDetection = T,
  humanDetection = T,
  landUseType = T,
  institutionLocality = T,
  geoOutliers = T,
  envOutliers = T,
  geoenvLowAccuracy = T
)
}
\value{
list with user modified settings
}
\description{
function used to select which types of tests you want in occTest workflow
}
\details{
See occTest::showTests for further information on tests used in the packages
}
\note{

}
\examples{
\dontrun{
example<-"goes here"
}
}
\references{

}
\seealso{

}
\author{
JM Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{}
\keyword{user}
