% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFuncs.R
\name{splitSpname}
\alias{splitSpname}
\title{Split joined species name}
\usage{
splitSpname(x)
}
\arguments{
\item{x}{species name join as a string separated by "_"}
}
\value{
a species name where genus and species are separated by a space
}
\description{
By default, it converts "Species_genus" format to "Species genus" format
}
\seealso{
Other internal: 
\code{\link{.get_os}()},
\code{\link{lazylogic}()},
\code{\link{rm.all}()}
}
\concept{internal}
