% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{timeAccuracy}
\alias{timeAccuracy}
\title{Flag occurrences with temporal inaccuracies}
\usage{
timeAccuracy(df, tf, iniTime = NA, endTime = NA, do = T, method = "all")
}
\arguments{
\item{df}{data.frame of species occurrences}

\item{tf}{character. column name in df containing the dataset with the date/time where the species is recorded}

\item{iniTime}{character or Date. indicates the initial time point considered in year-month-day format. Defaults to NA.}

\item{endTime}{character or Date. indicates the time point considered in year-month-day format. Defaults to NA.}

\item{do}{logical. Should tests be performed? Default TRUE}

\item{verbose}{logical. Print messages? Default FALSE}
}
\value{
data.frame
}
\description{
Detect records with low temporal accuracy
}
\details{
timeAccuracy function will implement different methods to assess the temporal scale of the parameters.\cr
}
\examples{
#see examples in vignetteXtra-occTest
}
\seealso{
Other analysis: 
\code{\link{.nearestcell3}()},
\code{\link{centroidDetection}()},
\code{\link{countryStatusRangeAnalysis}()},
\code{\link{duplicatesexcludeAnalysis}()},
\code{\link{geoEnvAccuracy}()},
\code{\link{humanDetection}()}
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{analysis}
\keyword{Analysis}
