% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{timeAccuracy}
\alias{timeAccuracy}
\title{Flag occurrences with temporal inaccuracies}
\usage{
timeAccuracy(
  df,
  tf,
  iniTime = NA,
  endTime = NA,
  do = T,
  method = "all",
  verbose = F
)
}
\arguments{
\item{df}{\emph{data.frame}. Dataset of species occurrences and their coordinates.}

\item{tf}{\emph{character}. Column name in df containing the dataset with the date/time where the species is recorded}

\item{iniTime}{\emph{character} or \emph{Date}. indicates the initial time point considered in year-month-day format. Defaults to NA.}

\item{endTime}{\emph{character} or \emph{Date}. indicates the time point considered in year-month-day format. Defaults to NA.}

\item{do}{\emph{logical}. Should tests be performed? Default TRUE}

\item{method}{\emph{character}. Vector of methods to be used. See details. Default 'all'}

\item{verbose}{\emph{logical}. Print messages? Default FALSE}
}
\value{
\emph{data.frame}
}
\description{
Detect records with low temporal accuracy
}
\details{
timeAccuracy function will implement different methods to assess the temporal scale of the parameters.\cr

timeAccuracy methods .\cr
Current implmented methods are:
'noDate' : tests for an existing value in the date field. \cr
'noDateFormatKnown' : tests for an known date format. \cr
'outDateRange' : tests for dates within the iniTime and endTime interval.
}
\examples{
\dontrun{
sp_file <- system.file('ext/exampleOccData.csv',package='occTest')
occ <- read.table(file = sp_file,header = T,sep = ',',as.is = T,nrows = 50)
timeAccuracy(df = occ,tf = 'DATE',
          iniTime = '2015-01-04',endTime = '2023-01-04')
#see examples in vignetteXtra-occTest
#library devtools()
#install_github ('vignetteXtra-occTest')
}
}
\seealso{
Other analysis: 
\code{\link{.nearestcell3}()},
\code{\link{centroidDetection}()},
\code{\link{countryStatusRangeAnalysis}()},
\code{\link{duplicatesexcludeAnalysis}()},
\code{\link{geoEnvAccuracy}()},
\code{\link{humanDetection}()}
}
\author{
Josep M Serra-Diaz (pep.serradiaz@agroparistech.fr)
}
\concept{analysis}
\keyword{Analysis}
